Specify Prior Distributions for SV Models

This function gives access to a larger set of prior distributions in case the default choice is unsatisfactory.

specify_priors(
  mu = sv_normal(mean = 0, sd = 100),
  phi = sv_beta(shape1 = 5, shape2 = 1.5),
  sigma2 = sv_gamma(shape = 0.5, rate = 0.5),
  nu = sv_infinity(),
  rho = sv_constant(0),
  latent0_variance = "stationary",
  beta = sv_multinormal(mean = 0, sd = 10000, dim = 1)
)

Arguments

mu
one of sv_normal or sv_constant

phi
one of sv_beta, sv_normal, or sv_constant. If sv_beta, then the specified beta distribution is the prior for (phi+1)/2

sigma2
one of sv_gamma, codesv_inverse_gamma, or sv_constant

nu
one of sv_infinity, sv_exponential, or sv_constant. If sv_exponential, then the specified exponential distribution is the prior for nu-2

rho
one of sv_beta or sv_constant. If sv_beta, then the specified beta distribution is the prior for (rho+1)/2

latent0_variance
either the character string "stationary" or an sv_constant object. If "stationary", then h0 ~ N(mu, sigma^2/(1-phi^2)). If an sv_constant object with value v, then h0 ~ N(mu, sigma^2/v). Here, N(b, B) stands for mean b and variance B

beta
an sv_multinormal object

See also