Prior Distributions in stochvol

The functions below can be supplied to specify_priors to overwrite the default set of prior distributions in svsample. The functions have mean, range, density, and print methods.

sv_constant(value)

sv_normal(mean = 0, sd = 1)

sv_multinormal(mean = 0, precision = NULL, sd = 1, dim = NA)

sv_gamma(shape, rate)

sv_inverse_gamma(shape, scale)

sv_beta(shape1, shape2)

sv_exponential(rate)

sv_infinity()

Arguments

value
The constant value for the degenerate constant distribution

mean
Expected value for the univariate normal distribution or mean vector of the multivariate normal distribution

sd
Standard deviation for the univariate normal distribution or constant scale of the multivariate normal distribution

precision
Precision matrix for the multivariate normal distribution

dim
(optional) Dimension of the multivariate distribution

shape
Shape parameter for the distribution

rate
Rate parameter for the distribution

scale
Scale parameter for the distribution

shape1
First shape parameter for the distribution

shape2
Second shape parameter for the distribution

Multivariate Normal

Multivariate normal objects can be specified several ways. The most general way is by calling sv_multinormal(mean, precision), which allows for arbitrary mean and (valid) precision arguments. Constant mean vectors and constant diagonal precision matrices of dimension D can be created two ways: either sv_multinormal(mean, sd, dim = D) or rep(sv_normal(mean, sd), length.out = D).

See also