Single MCMC update of Bayesian linear regression

Samples the coefficients of a linear regression. The prior distribution is multivariate normal and it is specified in prior_spec.

update_regressors(dependent_variable, independent_variables, beta, prior_spec)

Arguments

dependent_variable
the left hand side

independent_variables
the matrix of the independent variables. Has to be of same height as the length of the dependent variable

beta
the vector of the latent states used in MDA. Updated in place

prior_spec
prior specification object. See type_definitions.h

See also