Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean     SD Naive SE Time-series SE
#> mu    -11.4515 0.3665 0.008195        0.01533
#> phi     0.7693 0.1860 0.004158        0.02887
#> sigma   0.5150 0.1933 0.004323        0.02630
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> mu    -12.1443 -11.6281 -11.4563 -11.2706 -10.7537
#> phi     0.2269   0.6974   0.8179   0.9034   0.9762
#> sigma   0.2134   0.3667   0.4908   0.6364   0.9441
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>      -11.84567        0.79934        0.01787        0.03354 
#> 
#> 2. Quantiles for each variable:
#> 
#>   2.5%    25%    50%    75%  97.5% 
#> -13.38 -12.39 -11.88 -11.30 -10.25 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>         Mean     SD Naive SE Time-series SE
#> h_1   -11.91 0.7046  0.01576        0.03453
#> h_2   -12.01 0.6856  0.01533        0.03485
#> h_3   -12.26 0.7675  0.01716        0.04024
#> h_4   -12.37 0.7858  0.01757        0.04861
#> h_5   -12.38 0.7347  0.01643        0.04044
#> h_6   -12.43 0.8026  0.01795        0.04287
#> h_7   -12.37 0.7844  0.01754        0.04267
#> h_8   -12.22 0.7027  0.01571        0.03731
#> h_9   -12.13 0.7497  0.01676        0.04022
#> h_10  -11.93 0.7058  0.01578        0.03303
#> h_11  -11.55 0.5800  0.01297        0.03151
#> h_12  -11.64 0.6171  0.01380        0.02815
#> h_13  -11.69 0.6102  0.01364        0.02610
#> h_14  -11.74 0.6365  0.01423        0.02849
#> h_15  -11.72 0.6186  0.01383        0.02630
#> h_16  -11.53 0.5664  0.01266        0.03276
#> h_17  -11.65 0.5981  0.01337        0.02906
#> h_18  -11.70 0.6370  0.01424        0.03108
#> h_19  -11.78 0.6973  0.01559        0.03195
#> h_20  -11.71 0.6432  0.01438        0.02814
#> h_21  -11.61 0.5991  0.01340        0.02883
#> h_22  -11.66 0.5994  0.01340        0.02934
#> h_23  -11.70 0.6123  0.01369        0.02840
#> h_24  -11.91 0.6914  0.01546        0.03498
#> h_25  -11.99 0.7399  0.01654        0.04069
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>         2.5%    25%    50%    75%   97.5%
#> h_1   -13.17 -12.41 -11.94 -11.44 -10.478
#> h_2   -13.23 -12.50 -12.06 -11.54 -10.592
#> h_3   -13.80 -12.75 -12.26 -11.74 -10.752
#> h_4   -13.90 -12.88 -12.36 -11.85 -10.859
#> h_5   -13.81 -12.87 -12.36 -11.89 -10.915
#> h_6   -14.15 -12.92 -12.40 -11.91 -10.940
#> h_7   -14.07 -12.85 -12.35 -11.87 -10.876
#> h_8   -13.58 -12.67 -12.23 -11.77 -10.787
#> h_9   -13.69 -12.57 -12.11 -11.67 -10.707
#> h_10  -13.34 -12.35 -11.92 -11.53 -10.477
#> h_11  -12.56 -11.95 -11.58 -11.20 -10.251
#> h_12  -12.79 -12.05 -11.69 -11.28 -10.341
#> h_13  -12.86 -12.10 -11.71 -11.30 -10.457
#> h_14  -12.92 -12.14 -11.76 -11.34 -10.412
#> h_15  -12.93 -12.13 -11.75 -11.32 -10.447
#> h_16  -12.51 -11.93 -11.58 -11.18 -10.287
#> h_17  -12.79 -12.05 -11.66 -11.27 -10.410
#> h_18  -12.94 -12.10 -11.70 -11.33 -10.384
#> h_19  -13.15 -12.23 -11.77 -11.33 -10.351
#> h_20  -12.93 -12.14 -11.72 -11.31 -10.359
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean        SD  Naive SE Time-series SE
#> h_1   0.002759 0.0010409 2.328e-05      5.186e-05
#> h_2   0.002623 0.0009697 2.168e-05      4.806e-05
#> h_3   0.002346 0.0009249 2.068e-05      4.623e-05
#> h_4   0.002229 0.0008999 2.012e-05      5.456e-05
#> h_5   0.002198 0.0008434 1.886e-05      5.453e-05
#> h_6   0.002157 0.0008701 1.946e-05      4.642e-05
#> h_7   0.002218 0.0009013 2.015e-05      5.478e-05
#> h_8   0.002364 0.0008850 1.979e-05      5.982e-05
#> h_9   0.002491 0.0009665 2.161e-05      4.495e-05
#> h_10  0.002740 0.0010380 2.321e-05      4.683e-05
#> h_11  0.003248 0.0010439 2.334e-05      5.535e-05
#> h_12  0.003113 0.0010631 2.377e-05      4.800e-05
#> h_13  0.003031 0.0009834 2.199e-05      4.237e-05
#> h_14  0.002980 0.0010430 2.332e-05      4.620e-05
#> h_15  0.002994 0.0009740 2.178e-05      4.065e-05
#> h_16  0.003265 0.0009907 2.215e-05      5.769e-05
#> h_17  0.003090 0.0009648 2.157e-05      4.743e-05
#> h_18  0.003036 0.0010313 2.306e-05      4.923e-05
#> h_19  0.002947 0.0010928 2.444e-05      4.504e-05
#> h_20  0.003029 0.0010473 2.342e-05      4.380e-05
#> h_21  0.003149 0.0010165 2.273e-05      4.695e-05
#> h_22  0.003079 0.0009799 2.191e-05      4.668e-05
#> h_23  0.003022 0.0010346 2.314e-05      4.733e-05
#> h_24  0.002758 0.0010243 2.290e-05      4.746e-05
#> h_25  0.002657 0.0009914 2.217e-05      4.627e-05
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>            2.5%      25%      50%      75%    97.5%
#> h_1   0.0013809 0.002019 0.002551 0.003282 0.005305
#> h_2   0.0013387 0.001932 0.002409 0.003116 0.005011
#> h_3   0.0010058 0.001704 0.002173 0.002816 0.004626
#> h_4   0.0009607 0.001599 0.002071 0.002671 0.004386
#> h_5   0.0010006 0.001607 0.002071 0.002619 0.004264
#> h_6   0.0008474 0.001567 0.002027 0.002596 0.004211
#> h_7   0.0008796 0.001618 0.002083 0.002640 0.004348
#> h_8   0.0011265 0.001770 0.002210 0.002787 0.004546
#> h_9   0.0010623 0.001861 0.002346 0.002916 0.004731
#> h_10  0.0012707 0.002080 0.002575 0.003142 0.005309
#> h_11  0.0018717 0.002539 0.003055 0.003690 0.005943
#> h_12  0.0016738 0.002413 0.002901 0.003562 0.005681
#> h_13  0.0016114 0.002361 0.002866 0.003509 0.005362
#> h_14  0.0015622 0.002313 0.002792 0.003452 0.005485
#> h_15  0.0015608 0.002321 0.002816 0.003487 0.005389
#> h_16  0.0019161 0.002565 0.003065 0.003737 0.005838
#> h_17  0.0016681 0.002422 0.002939 0.003570 0.005490
#> h_18  0.0015498 0.002354 0.002881 0.003461 0.005560
#> h_19  0.0013960 0.002208 0.002776 0.003470 0.005652
#> h_20  0.0015599 0.002306 0.002851 0.003508 0.005631
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,] -1.834175e-04  1.919065e-01  3.537760e-02
#>     [2,]  8.346799e-05 -5.780474e-02 -1.871582e-02
#>     [3,]  3.946086e-06  7.613912e-03  2.154923e-02
#>     [4,]  4.199888e-05  1.126499e-02 -3.539692e-02
#>     [5,]  2.620620e-04  1.366091e-01  1.874739e-01
#>     [6,]  5.323612e-04  5.747054e-02  3.218873e-02
#>     [7,]  5.765312e-04  1.347622e-01  1.875007e-01
#>     [8,]  5.179180e-04  6.556952e-02  8.271941e-02
#>     [9,]  6.195452e-04 -7.267979e-03 -9.851478e-02
#>    [10,]  5.130271e-04  1.944362e-01  7.612757e-02
#>    [11,] -6.762959e-05  2.430460e-01  8.602107e-02
#>    [12,] -9.176114e-05  1.064139e-01 -8.660839e-02
#>    [13,] -2.038913e-05 -1.067040e-02  6.722786e-02
#>    [14,]  1.642984e-04 -5.955608e-02  1.365147e-01
#>    [15,]  2.767337e-04 -1.108548e-01  2.583301e-02
#>    [16,]  9.400213e-05  1.224767e-02  1.822903e-01
#>    [17,]  3.503369e-04 -5.903337e-02  1.703783e-01
#>    [18,]  1.445744e-04  1.001594e-02  9.167140e-02
#>    [19,]  1.893302e-05  1.231914e-01  8.441229e-02
#>    [20,]  1.154971e-04  2.650804e-02 -4.991165e-02
#>    [21,]  6.168802e-05  1.064268e-01  4.815879e-02
#>    [22,]  1.516808e-04  3.017920e-02 -7.899940e-02
#>    [23,] -1.851429e-04  9.617226e-02 -7.718960e-02
#>    [24,]  3.603448e-04 -1.324026e-01  1.366352e-01
#>    [25,]  5.861571e-04  9.854565e-02  3.406616e-02
#>    [26,]  9.948996e-06  6.612524e-02 -7.706607e-02
#>    [27,] -2.431813e-05  1.354963e-01  8.170537e-02
#>    [28,]  3.832356e-04  1.955025e-01  1.458943e-01
#>    [29,]  3.284449e-04 -6.168199e-02  1.236304e-01
#>    [30,]  3.278951e-04 -1.127620e-01 -3.986785e-02
#>    [31,] -1.566436e-06 -2.134772e-02  1.301189e-02
#>    [32,]  1.579203e-04  3.702340e-02 -3.998951e-02
#>    [33,] -1.643711e-04  4.978034e-02 -9.186108e-04
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>               tau_1      tau_2     tau_3     tau_4      tau_5      tau_6
#>              tau_7     tau_8      tau_9     tau_10    tau_11     tau_12
#>              tau_13     tau_14     tau_15     tau_16    tau_17     tau_18
#>              tau_19      tau_20     tau_21    tau_22     tau_23     tau_24
#>              tau_25     tau_26    tau_27     tau_28     tau_29     tau_30
#>              tau_31    tau_32     tau_33     tau_34    tau_35    tau_36
#>              tau_37    tau_38     tau_39     tau_40    tau_41     tau_42
#>             tau_43     tau_44    tau_45     tau_46     tau_47    tau_48
#>              tau_49     tau_50     tau_51     tau_52     tau_53      tau_54
#>              tau_55     tau_56     tau_57     tau_58    tau_59     tau_60
#>              tau_61    tau_62     tau_63     tau_64     tau_65    tau_66
#>              tau_67     tau_68     tau_69     tau_70     tau_71    tau_72
#>              tau_73     tau_74     tau_75     tau_76     tau_77     tau_78
#>              tau_79    tau_80    tau_81     tau_82    tau_83     tau_84
#>             tau_85     tau_86    tau_87     tau_88    tau_89     tau_90
#>              tau_91    tau_92     tau_93     tau_94     tau_95    tau_96
#>              tau_97    tau_98     tau_99   tau_100   tau_101   tau_102
#>             tau_103    tau_104    tau_105   tau_106   tau_107    tau_108
#>             tau_109   tau_110     tau_111   tau_112    tau_113    tau_114
#>            tau_115    tau_116    tau_117    tau_118   tau_119    tau_120
#>             tau_121   tau_122   tau_123    tau_124   tau_125    tau_126
#>             tau_127    tau_128    tau_129   tau_130    tau_131    tau_132
#>             tau_133   tau_134    tau_135    tau_136    tau_137   tau_138
#>             tau_139    tau_140    tau_141    tau_142    tau_143   tau_144
#>            tau_145    tau_146    tau_147    tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }