Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean      SD Naive SE Time-series SE
#> mu    -8.3708 0.69975 0.015647       0.032834
#> phi    0.9066 0.08482 0.001897       0.009092
#> sigma  0.3088 0.12674 0.002834       0.017298
#> 
#> 2. Quantiles for each variable:
#> 
#>          2.5%    25%     50%     75%   97.5%
#> mu    -9.2096 -8.692 -8.4714 -8.1964 -6.7885
#> phi    0.6680  0.882  0.9289  0.9623  0.9914
#> sigma  0.1296  0.221  0.2876  0.3732  0.6112
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>       -7.73546        0.64328        0.01438        0.03405 
#> 
#> 2. Quantiles for each variable:
#> 
#>   2.5%    25%    50%    75%  97.5% 
#> -8.981 -8.161 -7.765 -7.325 -6.422 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>         Mean     SD Naive SE Time-series SE
#> h_1   -7.673 0.5890 0.013171        0.02681
#> h_2   -7.574 0.5171 0.011563        0.02346
#> h_3   -7.462 0.4610 0.010309        0.02215
#> h_4   -7.527 0.4696 0.010500        0.02321
#> h_5   -7.654 0.5042 0.011275        0.02521
#> h_6   -7.755 0.5014 0.011212        0.02533
#> h_7   -7.800 0.4780 0.010689        0.02410
#> h_8   -7.869 0.4795 0.010722        0.02318
#> h_9   -8.004 0.5167 0.011553        0.02594
#> h_10  -8.108 0.5468 0.012227        0.03131
#> h_11  -8.180 0.5567 0.012449        0.03609
#> h_12  -8.207 0.5568 0.012450        0.03311
#> h_13  -8.183 0.5302 0.011855        0.02808
#> h_14  -8.107 0.5024 0.011234        0.02401
#> h_15  -7.961 0.4338 0.009700        0.01813
#> h_16  -8.071 0.4746 0.010612        0.01952
#> h_17  -8.163 0.5070 0.011337        0.02152
#> h_18  -8.177 0.5074 0.011346        0.02122
#> h_19  -8.116 0.4769 0.010663        0.01952
#> h_20  -8.046 0.4292 0.009596        0.01814
#> h_21  -8.164 0.4508 0.010081        0.01969
#> h_22  -8.253 0.4803 0.010740        0.02092
#> h_23  -8.269 0.4485 0.010028        0.01799
#> h_24  -8.322 0.4621 0.010333        0.01935
#> h_25  -8.364 0.4647 0.010390        0.01839
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>          2.5%    25%    50%    75%  97.5%
#> h_1    -8.768 -8.079 -7.700 -7.301 -6.436
#> h_2    -8.522 -7.931 -7.604 -7.256 -6.497
#> h_3    -8.255 -7.783 -7.495 -7.178 -6.490
#> h_4    -8.361 -7.852 -7.549 -7.238 -6.537
#> h_5    -8.630 -7.980 -7.684 -7.354 -6.584
#> h_6    -8.716 -8.081 -7.777 -7.445 -6.722
#> h_7    -8.671 -8.120 -7.820 -7.512 -6.752
#> h_8    -8.765 -8.179 -7.879 -7.575 -6.852
#> h_9    -9.009 -8.327 -8.004 -7.687 -6.987
#> h_10   -9.270 -8.437 -8.100 -7.762 -7.076
#> h_11   -9.335 -8.514 -8.167 -7.816 -7.124
#> h_12   -9.394 -8.544 -8.195 -7.845 -7.144
#> h_13   -9.306 -8.510 -8.182 -7.833 -7.229
#> h_14   -9.142 -8.410 -8.097 -7.804 -7.131
#> h_15   -8.726 -8.264 -7.980 -7.698 -7.071
#> h_16   -9.032 -8.363 -8.082 -7.770 -7.142
#> h_17   -9.260 -8.458 -8.141 -7.856 -7.219
#> h_18   -9.218 -8.475 -8.162 -7.865 -7.185
#> h_19   -9.093 -8.411 -8.118 -7.823 -7.111
#> h_20   -8.833 -8.338 -8.065 -7.767 -7.107
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean       SD  Naive SE Time-series SE
#> h_1   0.022546 0.007115 1.591e-04      0.0003179
#> h_2   0.023457 0.006527 1.460e-04      0.0002950
#> h_3   0.024639 0.006184 1.383e-04      0.0002959
#> h_4   0.023867 0.005959 1.333e-04      0.0002892
#> h_5   0.022496 0.005973 1.336e-04      0.0002932
#> h_6   0.021381 0.005656 1.265e-04      0.0002823
#> h_7   0.020847 0.005305 1.186e-04      0.0002597
#> h_8   0.020135 0.005099 1.140e-04      0.0002484
#> h_9   0.018911 0.005120 1.145e-04      0.0002529
#> h_10  0.018008 0.004989 1.116e-04      0.0002554
#> h_11  0.017390 0.004858 1.086e-04      0.0002825
#> h_12  0.017159 0.004800 1.073e-04      0.0002659
#> h_13  0.017296 0.004565 1.021e-04      0.0002292
#> h_14  0.017915 0.004620 1.033e-04      0.0001888
#> h_15  0.019143 0.004616 1.032e-04      0.0001907
#> h_16  0.018196 0.004584 1.025e-04      0.0001894
#> h_17  0.017438 0.004623 1.034e-04      0.0001897
#> h_18  0.017306 0.004447 9.943e-05      0.0001741
#> h_19  0.017790 0.004386 9.808e-05      0.0001754
#> h_20  0.018323 0.004137 9.250e-05      0.0001756
#> h_21  0.017314 0.004076 9.114e-05      0.0001833
#> h_22  0.016616 0.004198 9.387e-05      0.0001857
#> h_23  0.016426 0.003814 8.527e-05      0.0001539
#> h_24  0.016020 0.003789 8.471e-05      0.0001562
#> h_25  0.015688 0.003709 8.294e-05      0.0001469
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%   97.5%
#> h_1   0.012475 0.017606 0.021278 0.025981 0.04004
#> h_2   0.014107 0.018962 0.022331 0.026563 0.03883
#> h_3   0.016126 0.020414 0.023576 0.027621 0.03896
#> h_4   0.015289 0.019718 0.022947 0.026810 0.03807
#> h_5   0.013367 0.018503 0.021454 0.025301 0.03717
#> h_6   0.012804 0.017592 0.020476 0.024179 0.03471
#> h_7   0.013098 0.017250 0.020038 0.023375 0.03419
#> h_8   0.012497 0.016747 0.019454 0.022651 0.03252
#> h_9   0.011061 0.015552 0.018277 0.021415 0.03040
#> h_10  0.009706 0.014720 0.017425 0.020631 0.02907
#> h_11  0.009398 0.014164 0.016848 0.020082 0.02839
#> h_12  0.009125 0.013956 0.016615 0.019789 0.02810
#> h_13  0.009535 0.014196 0.016726 0.019907 0.02693
#> h_14  0.010349 0.014921 0.017448 0.020201 0.02828
#> h_15  0.012737 0.016050 0.018503 0.021306 0.02914
#> h_16  0.010935 0.015279 0.017580 0.020545 0.02813
#> h_17  0.009752 0.014567 0.017068 0.019685 0.02707
#> h_18  0.009960 0.014445 0.016887 0.019598 0.02753
#> h_19  0.010606 0.014911 0.017267 0.020008 0.02857
#> h_20  0.012077 0.015467 0.017727 0.020577 0.02863
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,]  8.292107e-04 -1.708473e-01  4.127888e-02
#>     [2,]  1.063175e-03 -2.347933e-01  8.484073e-02
#>     [3,] -1.560377e-03 -1.365503e-01  2.334856e-01
#>     [4,] -1.719833e-03 -4.990347e-02  5.980720e-02
#>     [5,] -5.119081e-04 -6.442167e-02  1.411062e-01
#>     [6,] -3.224080e-04 -1.019533e-01  3.367425e-02
#>     [7,]  1.334531e-04 -2.347581e-01  1.853540e-01
#>     [8,] -9.652637e-04 -3.003606e-02  1.203787e-01
#>     [9,] -9.962213e-05 -1.269651e-01 -1.032983e-02
#>    [10,]  1.546931e-03 -6.801726e-02 -4.492254e-02
#>    [11,]  2.661069e-03 -1.061335e-01  1.148075e-01
#>    [12,]  8.338631e-04 -1.063356e-01  3.101249e-02
#>    [13,]  1.710662e-03 -1.465339e-01  1.539371e-03
#>    [14,]  9.604437e-04 -9.062632e-02 -2.965592e-02
#>    [15,]  6.791585e-04 -1.615827e-01  3.849629e-02
#>    [16,]  7.221096e-04 -2.008197e-01  1.404326e-01
#>    [17,]  1.014100e-03 -2.572823e-01  1.036051e-01
#>    [18,]  8.322912e-04 -1.441707e-01  1.241591e-01
#>    [19,]  5.555431e-04  8.079878e-02  1.498343e-01
#>    [20,] -8.241170e-04 -1.609456e-01  1.460530e-01
#>    [21,] -2.410000e-05 -1.505330e-01  2.064093e-01
#>    [22,] -1.890707e-03 -2.312099e-01  3.039760e-02
#>    [23,] -1.586933e-03 -1.859075e-01  3.013809e-01
#>    [24,]  5.956256e-04 -2.407225e-01  7.000604e-02
#>    [25,] -3.484256e-04 -8.490871e-02 -7.228425e-03
#>    [26,] -7.128661e-04 -1.394403e-01  4.087578e-02
#>    [27,] -6.754886e-04 -1.644249e-01  1.503773e-01
#>    [28,]  1.474838e-04 -1.814776e-01  9.903247e-02
#>    [29,] -5.588502e-04 -9.164008e-02  1.420768e-01
#>    [30,] -1.846495e-03  2.263545e-02 -4.706409e-02
#>    [31,] -2.121471e-03 -1.203898e-01  1.235486e-01
#>    [32,] -5.986677e-04  3.963488e-02 -1.197843e-02
#>    [33,]  7.585001e-04  1.022006e-02  4.273002e-02
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>               tau_1       tau_2     tau_3      tau_4      tau_5      tau_6
#>               tau_7      tau_8      tau_9     tau_10    tau_11    tau_12
#>              tau_13    tau_14     tau_15     tau_16     tau_17     tau_18
#>             tau_19     tau_20     tau_21     tau_22     tau_23      tau_24
#>              tau_25     tau_26     tau_27    tau_28     tau_29     tau_30
#>              tau_31    tau_32     tau_33     tau_34     tau_35     tau_36
#>             tau_37     tau_38     tau_39     tau_40     tau_41     tau_42
#>              tau_43     tau_44     tau_45     tau_46     tau_47     tau_48
#>              tau_49    tau_50     tau_51    tau_52     tau_53     tau_54
#>              tau_55      tau_56     tau_57     tau_58     tau_59    tau_60
#>             tau_61     tau_62     tau_63    tau_64     tau_65    tau_66
#>              tau_67     tau_68     tau_69     tau_70    tau_71    tau_72
#>              tau_73     tau_74      tau_75     tau_76     tau_77     tau_78
#>              tau_79     tau_80     tau_81    tau_82     tau_83     tau_84
#>              tau_85     tau_86     tau_87     tau_88     tau_89     tau_90
#>              tau_91     tau_92     tau_93     tau_94     tau_95     tau_96
#>              tau_97     tau_98    tau_99     tau_100    tau_101    tau_102
#>             tau_103    tau_104    tau_105   tau_106    tau_107   tau_108
#>            tau_109   tau_110    tau_111    tau_112    tau_113    tau_114
#>             tau_115    tau_116    tau_117   tau_118    tau_119    tau_120
#>            tau_121    tau_122    tau_123    tau_124    tau_125    tau_126
#>             tau_127   tau_128    tau_129    tau_130   tau_131    tau_132
#>             tau_133    tau_134    tau_135   tau_136    tau_137   tau_138
#>            tau_139    tau_140   tau_141    tau_142    tau_143    tau_144
#>             tau_145    tau_146    tau_147    tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }