Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> mu    -9.9977 0.1655 0.003701       0.009674
#> phi    0.5758 0.2694 0.006023       0.038287
#> sigma  0.2914 0.1408 0.003148       0.012505
#> 
#> 2. Quantiles for each variable:
#> 
#>            2.5%      25%     50%     75%   97.5%
#> mu    -10.32815 -10.1020 -9.9961 -9.8889 -9.6883
#> phi    -0.05249   0.4186  0.6321  0.7859  0.9285
#> sigma   0.02505   0.2004  0.2889  0.3833  0.5752
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>      -9.933768       0.434301       0.009711       0.013150 
#> 
#> 2. Quantiles for each variable:
#> 
#>    2.5%     25%     50%     75%   97.5% 
#> -10.792 -10.188  -9.929  -9.690  -9.015 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> h_1    -9.892 0.3939 0.008807        0.01336
#> h_2    -9.942 0.4112 0.009195        0.01388
#> h_3    -9.920 0.3983 0.008906        0.01263
#> h_4    -9.816 0.3476 0.007773        0.01188
#> h_5    -9.891 0.3622 0.008099        0.01249
#> h_6    -9.954 0.3721 0.008320        0.01372
#> h_7    -9.996 0.4103 0.009175        0.01459
#> h_8   -10.030 0.4124 0.009223        0.01463
#> h_9    -9.999 0.3990 0.008921        0.01370
#> h_10   -9.976 0.3962 0.008858        0.01413
#> h_11   -9.924 0.3653 0.008169        0.01325
#> h_12  -10.044 0.3949 0.008830        0.01384
#> h_13  -10.092 0.4246 0.009494        0.01657
#> h_14  -10.089 0.4028 0.009008        0.01794
#> h_15  -10.122 0.4288 0.009587        0.01847
#> h_16  -10.103 0.4345 0.009715        0.01614
#> h_17  -10.066 0.3967 0.008870        0.01583
#> h_18  -10.100 0.4189 0.009368        0.01643
#> h_19  -10.073 0.4047 0.009050        0.01603
#> h_20  -10.084 0.4074 0.009109        0.01436
#> h_21  -10.064 0.3936 0.008801        0.01522
#> h_22   -9.963 0.3709 0.008294        0.01623
#> h_23  -10.062 0.3991 0.008925        0.01633
#> h_24  -10.147 0.4336 0.009695        0.02144
#> h_25  -10.200 0.4444 0.009937        0.02505
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>         2.5%     25%     50%    75%  97.5%
#> h_1   -10.64 -10.129  -9.905 -9.671 -9.041
#> h_2   -10.78 -10.176  -9.941 -9.722 -9.068
#> h_3   -10.71 -10.152  -9.924 -9.703 -9.094
#> h_4   -10.44 -10.034  -9.850 -9.616 -9.063
#> h_5   -10.60 -10.104  -9.902 -9.682 -9.127
#> h_6   -10.70 -10.179  -9.952 -9.743 -9.188
#> h_7   -10.81 -10.224  -9.983 -9.771 -9.198
#> h_8   -10.89 -10.284 -10.008 -9.780 -9.205
#> h_9   -10.83 -10.234  -9.983 -9.766 -9.192
#> h_10  -10.75 -10.193  -9.976 -9.753 -9.176
#> h_11  -10.63 -10.154  -9.932 -9.718 -9.118
#> h_12  -10.87 -10.288 -10.028 -9.807 -9.259
#> h_13  -11.01 -10.347 -10.057 -9.833 -9.274
#> h_14  -10.96 -10.328 -10.053 -9.841 -9.305
#> h_15  -11.09 -10.362 -10.084 -9.844 -9.325
#> h_16  -11.02 -10.350 -10.063 -9.838 -9.287
#> h_17  -10.94 -10.294 -10.033 -9.814 -9.301
#> h_18  -10.98 -10.341 -10.077 -9.849 -9.289
#> h_19  -10.92 -10.313 -10.053 -9.823 -9.312
#> h_20  -10.94 -10.312 -10.057 -9.832 -9.298
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean       SD  Naive SE Time-series SE
#> h_1   0.007253 0.001505 3.366e-05      5.306e-05
#> h_2   0.007085 0.001502 3.358e-05      5.133e-05
#> h_3   0.007157 0.001482 3.313e-05      4.723e-05
#> h_4   0.007502 0.001373 3.071e-05      4.684e-05
#> h_5   0.007234 0.001358 3.037e-05      4.714e-05
#> h_6   0.007015 0.001347 3.011e-05      4.865e-05
#> h_7   0.006897 0.001469 3.285e-05      4.732e-05
#> h_8   0.006779 0.001422 3.179e-05      4.891e-05
#> h_9   0.006877 0.001397 3.123e-05      4.745e-05
#> h_10  0.006956 0.001433 3.205e-05      4.965e-05
#> h_11  0.007121 0.001373 3.071e-05      4.904e-05
#> h_12  0.006719 0.001345 3.007e-05      4.556e-05
#> h_13  0.006581 0.001404 3.138e-05      5.247e-05
#> h_14  0.006574 0.001322 2.955e-05      5.777e-05
#> h_15  0.006482 0.001352 3.023e-05      5.562e-05
#> h_16  0.006549 0.001405 3.142e-05      5.058e-05
#> h_17  0.006648 0.001300 2.908e-05      4.992e-05
#> h_18  0.006550 0.001377 3.080e-05      5.202e-05
#> h_19  0.006629 0.001354 3.027e-05      5.224e-05
#> h_20  0.006595 0.001341 2.998e-05      4.733e-05
#> h_21  0.006652 0.001313 2.936e-05      4.951e-05
#> h_22  0.006986 0.001344 3.004e-05      5.820e-05
#> h_23  0.006663 0.001343 3.004e-05      5.316e-05
#> h_24  0.006407 0.001374 3.072e-05      6.369e-05
#> h_25  0.006243 0.001352 3.023e-05      7.066e-05
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.004893 0.006316 0.007064 0.007941 0.010883
#> h_2   0.004565 0.006170 0.006940 0.007741 0.010738
#> h_3   0.004713 0.006245 0.007001 0.007815 0.010600
#> h_4   0.005412 0.006625 0.007263 0.008164 0.010766
#> h_5   0.004986 0.006397 0.007075 0.007898 0.010425
#> h_6   0.004747 0.006161 0.006903 0.007663 0.010112
#> h_7   0.004503 0.006025 0.006795 0.007556 0.010061
#> h_8   0.004320 0.005845 0.006711 0.007523 0.010024
#> h_9   0.004453 0.005993 0.006795 0.007573 0.010092
#> h_10  0.004635 0.006117 0.006819 0.007624 0.010173
#> h_11  0.004920 0.006240 0.006972 0.007758 0.010473
#> h_12  0.004371 0.005834 0.006643 0.007422 0.009761
#> h_13  0.004066 0.005666 0.006547 0.007326 0.009689
#> h_14  0.004160 0.005719 0.006562 0.007297 0.009537
#> h_15  0.003906 0.005624 0.006459 0.007286 0.009443
#> h_16  0.004039 0.005655 0.006530 0.007307 0.009624
#> h_17  0.004202 0.005818 0.006627 0.007395 0.009558
#> h_18  0.004127 0.005683 0.006483 0.007265 0.009616
#> h_19  0.004262 0.005763 0.006562 0.007363 0.009507
#> h_20  0.004218 0.005766 0.006547 0.007330 0.009570
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,] -1.102729e-03  2.036801e-01 -4.780464e-04
#>     [2,] -5.575007e-04  1.752961e-01 -3.747227e-02
#>     [3,]  3.733575e-04 -1.650824e-02 -1.412445e-02
#>     [4,] -2.939817e-05  8.573432e-02  6.166771e-02
#>     [5,] -6.323833e-04  1.522970e-01 -2.277825e-01
#>     [6,] -8.279000e-04  2.373854e-01 -7.067195e-02
#>     [7,] -8.968188e-04  2.894781e-01 -2.423491e-01
#>     [8,] -4.095486e-04  1.488230e-01 -2.348768e-01
#>     [9,]  1.198898e-04  1.693641e-01 -1.004641e-01
#>    [10,] -2.918384e-04  2.093653e-01 -7.741129e-02
#>    [11,]  2.581632e-04  3.157770e-01 -1.328836e-02
#>    [12,] -4.113935e-04  2.407192e-01 -3.793457e-02
#>    [13,] -9.737893e-04  9.600889e-02 -2.986249e-02
#>    [14,]  1.209724e-04  9.062931e-02 -1.217411e-01
#>    [15,]  1.376857e-04  2.421994e-01 -2.405957e-01
#>    [16,] -4.937391e-04  1.927745e-01 -1.329328e-01
#>    [17,] -8.720469e-05  5.561517e-02 -1.325526e-01
#>    [18,] -3.099370e-04  6.746871e-02 -1.742432e-01
#>    [19,] -9.748621e-04  4.668080e-03 -5.716450e-02
#>    [20,] -3.257796e-04  1.607068e-01 -1.839343e-01
#>    [21,] -4.563198e-04  2.651057e-01 -2.045838e-01
#>    [22,] -1.887876e-03  2.250794e-01 -3.853647e-01
#>    [23,] -1.166686e-03  2.487582e-01 -1.013036e-01
#>    [24,] -9.606880e-04  1.710472e-01 -3.300460e-02
#>    [25,] -6.595042e-04  2.215730e-01 -1.057468e-01
#>    [26,] -1.364495e-04  2.604809e-01 -1.718039e-01
#>    [27,]  2.638001e-04  1.499939e-01 -2.104470e-01
#>    [28,] -9.547951e-04  7.144388e-02 -4.872984e-02
#>    [29,] -8.875535e-04  1.164439e-01 -1.148565e-01
#>    [30,] -9.587833e-04  1.416373e-01 -1.103098e-01
#>    [31,] -9.744139e-04  1.739407e-01 -2.976351e-01
#>    [32,] -4.866367e-04  1.161576e-01 -1.818575e-01
#>    [33,]  2.743538e-04  9.721573e-02 -1.335829e-01
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>              tau_1      tau_2      tau_3     tau_4     tau_5     tau_6
#>               tau_7     tau_8      tau_9    tau_10    tau_11     tau_12
#>             tau_13     tau_14    tau_15    tau_16    tau_17    tau_18    tau_19
#>              tau_20    tau_21    tau_22     tau_23     tau_24    tau_25
#>             tau_26     tau_27    tau_28    tau_29    tau_30    tau_31    tau_32
#>             tau_33    tau_34    tau_35    tau_36    tau_37    tau_38     tau_39
#>              tau_40    tau_41    tau_42    tau_43    tau_44    tau_45    tau_46
#>             tau_47    tau_48    tau_49    tau_50     tau_51     tau_52
#>              tau_53    tau_54    tau_55    tau_56    tau_57    tau_58
#>              tau_59     tau_60     tau_61     tau_62    tau_63    tau_64
#>             tau_65     tau_66    tau_67     tau_68     tau_69    tau_70
#>             tau_71     tau_72     tau_73     tau_74     tau_75    tau_76
#>              tau_77    tau_78    tau_79    tau_80     tau_81    tau_82
#>             tau_83     tau_84    tau_85     tau_86    tau_87    tau_88
#>             tau_89     tau_90     tau_91     tau_92     tau_93    tau_94
#>             tau_95     tau_96     tau_97    tau_98    tau_99   tau_100
#>            tau_101    tau_102   tau_103    tau_104    tau_105   tau_106
#>            tau_107    tau_108   tau_109    tau_110   tau_111   tau_112
#>             tau_113   tau_114   tau_115   tau_116   tau_117   tau_118
#>             tau_119   tau_120   tau_121    tau_122    tau_123   tau_124
#>             tau_125    tau_126    tau_127    tau_128    tau_129   tau_130
#>            tau_131   tau_132   tau_133   tau_134   tau_135    tau_136   tau_137
#>            tau_138    tau_139    tau_140   tau_141   tau_142   tau_143
#>             tau_144   tau_145   tau_146   tau_147   tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }