Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean     SD Naive SE Time-series SE
#> mu    -10.5431 0.5000 0.011179        0.02184
#> phi     0.8538 0.1295 0.002895        0.02088
#> sigma   0.4637 0.2041 0.004564        0.03640
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%   97.5%
#> mu    -11.3883 -10.7838 -10.5568 -10.3119 -9.5206
#> phi     0.4469   0.8095   0.8912   0.9420  0.9870
#> sigma   0.1824   0.3153   0.4228   0.5705  0.9695
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>       -9.81794        0.76875        0.01719        0.02777 
#> 
#> 2. Quantiles for each variable:
#> 
#>    2.5%     25%     50%     75%   97.5% 
#> -11.307 -10.308  -9.835  -9.334  -8.235 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> h_1    -9.696 0.6151  0.01375        0.02636
#> h_2    -9.692 0.5789  0.01295        0.02557
#> h_3    -9.629 0.5337  0.01193        0.02296
#> h_4    -9.967 0.6871  0.01536        0.04502
#> h_5   -10.118 0.7350  0.01643        0.06002
#> h_6   -10.109 0.7082  0.01584        0.04893
#> h_7    -9.960 0.5628  0.01259        0.02530
#> h_8   -10.089 0.6887  0.01540        0.06279
#> h_9   -10.068 0.6906  0.01544        0.07234
#> h_10   -9.889 0.5903  0.01320        0.02828
#> h_11   -9.804 0.6102  0.01364        0.02728
#> h_12   -9.536 0.5309  0.01187        0.02338
#> h_13   -9.512 0.5832  0.01304        0.02428
#> h_14   -9.403 0.5858  0.01310        0.02465
#> h_15   -9.110 0.5272  0.01179        0.02674
#> h_16   -9.310 0.5449  0.01218        0.02278
#> h_17   -9.395 0.5430  0.01214        0.02262
#> h_18   -9.348 0.5005  0.01119        0.02219
#> h_19   -9.779 0.6145  0.01374        0.02794
#> h_20  -10.130 0.7314  0.01636        0.05181
#> h_21  -10.290 0.7104  0.01589        0.05261
#> h_22  -10.346 0.6201  0.01387        0.04093
#> h_23  -10.584 0.7360  0.01646        0.07049
#> h_24  -10.671 0.7375  0.01649        0.05677
#> h_25  -10.640 0.7171  0.01603        0.04504
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>         2.5%     25%     50%     75%   97.5%
#> h_1   -10.80 -10.100  -9.740  -9.315  -8.354
#> h_2   -10.76 -10.079  -9.729  -9.328  -8.443
#> h_3   -10.57  -9.996  -9.680  -9.269  -8.519
#> h_4   -11.43 -10.366  -9.942  -9.538  -8.733
#> h_5   -11.75 -10.532 -10.078  -9.637  -8.775
#> h_6   -11.66 -10.524 -10.078  -9.641  -8.796
#> h_7   -11.00 -10.359  -9.984  -9.605  -8.782
#> h_8   -11.62 -10.467 -10.052  -9.655  -8.805
#> h_9   -11.49 -10.476 -10.029  -9.633  -8.813
#> h_10  -11.09 -10.271  -9.885  -9.502  -8.710
#> h_11  -11.02 -10.186  -9.802  -9.422  -8.618
#> h_12  -10.45  -9.897  -9.569  -9.222  -8.310
#> h_13  -10.62  -9.886  -9.525  -9.163  -8.296
#> h_14  -10.55  -9.761  -9.418  -9.059  -8.170
#> h_15  -10.02  -9.478  -9.160  -8.804  -7.917
#> h_16  -10.30  -9.680  -9.340  -8.974  -8.198
#> h_17  -10.42  -9.746  -9.420  -9.055  -8.216
#> h_18  -10.21  -9.700  -9.385  -9.048  -8.213
#> h_19  -11.11 -10.159  -9.765  -9.392  -8.577
#> h_20  -11.84 -10.534 -10.084  -9.632  -8.855
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean        SD  Naive SE Time-series SE
#> h_1   0.008243 0.0028755 6.430e-05      1.217e-04
#> h_2   0.008207 0.0025716 5.750e-05      1.124e-04
#> h_3   0.008413 0.0023896 5.343e-05      1.043e-04
#> h_4   0.007245 0.0024036 5.375e-05      1.391e-04
#> h_5   0.006774 0.0024162 5.403e-05      1.590e-04
#> h_6   0.006777 0.0023422 5.237e-05      1.178e-04
#> h_7   0.007156 0.0021148 4.729e-05      9.495e-05
#> h_8   0.006823 0.0023137 5.174e-05      1.384e-04
#> h_9   0.006894 0.0023332 5.217e-05      1.143e-04
#> h_10  0.007440 0.0022491 5.029e-05      1.011e-04
#> h_11  0.007781 0.0024151 5.400e-05      1.050e-04
#> h_12  0.008817 0.0025922 5.796e-05      1.135e-04
#> h_13  0.008983 0.0028337 6.336e-05      1.136e-04
#> h_14  0.009487 0.0029347 6.562e-05      1.241e-04
#> h_15  0.010908 0.0032194 7.199e-05      1.683e-04
#> h_16  0.009883 0.0028879 6.458e-05      1.079e-04
#> h_17  0.009472 0.0027636 6.180e-05      1.146e-04
#> h_18  0.009643 0.0026085 5.833e-05      1.188e-04
#> h_19  0.007887 0.0024907 5.569e-05      1.119e-04
#> h_20  0.006722 0.0023417 5.236e-05      1.203e-04
#> h_21  0.006193 0.0021647 4.840e-05      1.169e-04
#> h_22  0.005944 0.0018741 4.191e-05      1.150e-04
#> h_23  0.005368 0.0019223 4.298e-05      1.122e-04
#> h_24  0.005143 0.0018416 4.118e-05      1.236e-04
#> h_25  0.005210 0.0018673 4.175e-05      1.130e-04
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.004520 0.006411 0.007674 0.009489 0.015348
#> h_2   0.004600 0.006476 0.007717 0.009429 0.014677
#> h_3   0.005067 0.006752 0.007908 0.009710 0.014126
#> h_4   0.003291 0.005610 0.006938 0.008487 0.012698
#> h_5   0.002804 0.005163 0.006480 0.008079 0.012435
#> h_6   0.002938 0.005185 0.006480 0.008064 0.012305
#> h_7   0.004087 0.005631 0.006793 0.008210 0.012391
#> h_8   0.002998 0.005335 0.006565 0.008005 0.012248
#> h_9   0.003193 0.005312 0.006641 0.008094 0.012199
#> h_10  0.003911 0.005885 0.007137 0.008643 0.012843
#> h_11  0.004040 0.006140 0.007438 0.008996 0.013446
#> h_12  0.005384 0.007095 0.008357 0.009940 0.015686
#> h_13  0.004937 0.007135 0.008544 0.010242 0.015795
#> h_14  0.005111 0.007592 0.009015 0.010786 0.016824
#> h_15  0.006672 0.008749 0.010257 0.012256 0.019096
#> h_16  0.005787 0.007908 0.009374 0.011256 0.016587
#> h_17  0.005472 0.007649 0.009004 0.010806 0.016439
#> h_18  0.006063 0.007830 0.009163 0.010847 0.016462
#> h_19  0.003873 0.006224 0.007578 0.009133 0.013724
#> h_20  0.002687 0.005160 0.006460 0.008097 0.011942
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,]  6.217202e-04 -1.406193e-01  6.060663e-02
#>     [2,]  4.003415e-04 -2.078121e-01  6.740864e-02
#>     [3,]  9.020988e-04 -2.880491e-01 -5.059775e-02
#>     [4,]  1.252596e-03 -2.100334e-01 -4.936486e-02
#>     [5,]  8.918388e-04 -2.383385e-01 -3.414578e-03
#>     [6,]  1.019466e-03 -7.403813e-02 -6.965792e-02
#>     [7,]  1.263879e-03 -7.093539e-02  4.036089e-03
#>     [8,]  3.063550e-04 -1.677492e-01 -1.566498e-01
#>     [9,]  3.849175e-04 -8.750351e-02  7.395893e-02
#>    [10,]  8.216221e-04 -1.152725e-01 -1.237276e-01
#>    [11,]  4.971096e-04 -1.241966e-01 -6.828675e-02
#>    [12,]  6.450314e-04 -2.519909e-01  2.003696e-02
#>    [13,]  2.469491e-04 -2.787322e-01 -8.121211e-02
#>    [14,]  1.077216e-03 -1.567350e-01 -7.711840e-02
#>    [15,]  8.848478e-04 -1.430862e-01 -3.559804e-02
#>    [16,]  4.166115e-04 -5.160896e-02  2.384901e-02
#>    [17,]  6.724294e-04 -4.959878e-02  2.622912e-02
#>    [18,]  8.139108e-04 -1.038423e-01  1.234020e-02
#>    [19,]  1.695884e-04 -1.004706e-01 -5.837611e-02
#>    [20,]  4.996997e-04 -6.637856e-02 -3.946809e-02
#>    [21,]  6.570966e-04  3.315653e-02 -1.822895e-01
#>    [22,]  5.158783e-04  1.205062e-01 -8.708718e-02
#>    [23,]  7.127173e-04  5.598584e-03 -5.123399e-03
#>    [24,]  5.496548e-04  5.997469e-02 -2.794178e-02
#>    [25,]  6.284336e-04 -2.774966e-02 -1.103061e-01
#>    [26,]  9.154133e-04 -2.471335e-02 -2.442683e-01
#>    [27,]  6.607042e-04 -4.881368e-02 -2.339453e-01
#>    [28,]  4.473476e-04 -7.889161e-02 -2.326438e-01
#>    [29,]  2.132297e-04 -5.985739e-02 -1.642597e-01
#>    [30,]  3.880287e-04 -1.502644e-01 -1.022674e-01
#>    [31,]  8.983072e-04 -1.110068e-01 -7.507777e-02
#>    [32,]  9.253463e-04 -3.043799e-01 -9.058288e-02
#>    [33,]  6.431138e-04 -7.748809e-02 -1.070090e-01
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>               tau_1      tau_2      tau_3     tau_4      tau_5     tau_6
#>               tau_7      tau_8       tau_9     tau_10     tau_11     tau_12
#>              tau_13     tau_14     tau_15     tau_16     tau_17      tau_18
#>               tau_19    tau_20      tau_21    tau_22     tau_23     tau_24
#>              tau_25      tau_26     tau_27     tau_28      tau_29     tau_30
#>              tau_31     tau_32     tau_33     tau_34      tau_35     tau_36
#>              tau_37     tau_38    tau_39     tau_40     tau_41    tau_42
#>              tau_43     tau_44     tau_45     tau_46     tau_47     tau_48
#>              tau_49     tau_50    tau_51     tau_52     tau_53     tau_54
#>              tau_55     tau_56     tau_57     tau_58    tau_59     tau_60
#>              tau_61     tau_62     tau_63      tau_64     tau_65     tau_66
#>              tau_67     tau_68      tau_69     tau_70      tau_71    tau_72
#>              tau_73    tau_74     tau_75     tau_76     tau_77     tau_78
#>               tau_79     tau_80     tau_81    tau_82     tau_83      tau_84
#>              tau_85      tau_86     tau_87     tau_88     tau_89     tau_90
#>              tau_91    tau_92     tau_93     tau_94     tau_95     tau_96
#>              tau_97     tau_98      tau_99    tau_100    tau_101    tau_102
#>              tau_103     tau_104    tau_105     tau_106    tau_107    tau_108
#>              tau_109     tau_110     tau_111    tau_112    tau_113    tau_114
#>             tau_115   tau_116    tau_117    tau_118   tau_119   tau_120
#>             tau_121     tau_122    tau_123    tau_124    tau_125   tau_126
#>              tau_127     tau_128    tau_129    tau_130    tau_131    tau_132
#>             tau_133    tau_134     tau_135     tau_136    tau_137    tau_138
#>             tau_139    tau_140   tau_141    tau_142   tau_143    tau_144
#>             tau_145    tau_146    tau_147    tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }