Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean      SD  Naive SE Time-series SE
#> mu    -11.8309 0.79116 0.0176909       0.026575
#> phi     0.9532 0.03615 0.0008083       0.003743
#> sigma   0.2759 0.09924 0.0022191       0.012891
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> mu    -13.2543 -12.1758 -11.8489 -11.4970 -10.2654
#> phi     0.8600   0.9365   0.9624   0.9789   0.9943
#> sigma   0.1288   0.1971   0.2642   0.3390   0.4999
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>      -12.28478        0.66745        0.01492        0.02856 
#> 
#> 2. Quantiles for each variable:
#> 
#>   2.5%    25%    50%    75%  97.5% 
#> -13.65 -12.71 -12.27 -11.86 -11.04 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>         Mean     SD Naive SE Time-series SE
#> h_1   -12.31 0.6262 0.014002        0.02912
#> h_2   -12.28 0.5738 0.012830        0.02555
#> h_3   -12.24 0.5404 0.012084        0.02487
#> h_4   -12.24 0.5454 0.012195        0.02531
#> h_5   -12.20 0.5234 0.011705        0.02360
#> h_6   -12.12 0.4858 0.010862        0.02180
#> h_7   -12.01 0.4323 0.009666        0.01972
#> h_8   -11.95 0.4220 0.009437        0.01973
#> h_9   -11.95 0.4270 0.009548        0.02132
#> h_10  -11.92 0.4087 0.009140        0.01979
#> h_11  -11.99 0.4085 0.009134        0.01787
#> h_12  -12.18 0.4535 0.010141        0.01714
#> h_13  -12.32 0.4862 0.010872        0.02040
#> h_14  -12.44 0.5077 0.011353        0.02185
#> h_15  -12.50 0.5109 0.011424        0.02422
#> h_16  -12.53 0.4893 0.010941        0.02443
#> h_17  -12.60 0.4913 0.010987        0.02207
#> h_18  -12.66 0.5028 0.011243        0.02408
#> h_19  -12.70 0.4958 0.011086        0.02175
#> h_20  -12.68 0.4884 0.010921        0.02157
#> h_21  -12.68 0.4770 0.010666        0.02112
#> h_22  -12.64 0.4428 0.009902        0.02116
#> h_23  -12.62 0.4325 0.009670        0.02225
#> h_24  -12.67 0.4391 0.009819        0.02098
#> h_25  -12.76 0.4719 0.010551        0.01986
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>         2.5%    25%    50%    75%   97.5%
#> h_1   -13.59 -12.71 -12.31 -11.91 -11.083
#> h_2   -13.38 -12.66 -12.29 -11.92 -11.121
#> h_3   -13.27 -12.62 -12.26 -11.90 -11.143
#> h_4   -13.29 -12.61 -12.24 -11.88 -11.157
#> h_5   -13.26 -12.55 -12.20 -11.87 -11.152
#> h_6   -13.08 -12.44 -12.13 -11.80 -11.134
#> h_7   -12.78 -12.32 -12.02 -11.74 -11.104
#> h_8   -12.74 -12.24 -11.98 -11.68 -11.061
#> h_9   -12.76 -12.24 -11.96 -11.68 -11.056
#> h_10  -12.67 -12.21 -11.93 -11.66 -11.062
#> h_11  -12.79 -12.27 -12.00 -11.74 -11.167
#> h_12  -13.07 -12.46 -12.17 -11.89 -11.250
#> h_13  -13.30 -12.64 -12.32 -12.01 -11.400
#> h_14  -13.47 -12.75 -12.44 -12.11 -11.480
#> h_15  -13.58 -12.82 -12.49 -12.19 -11.526
#> h_16  -13.49 -12.86 -12.54 -12.23 -11.552
#> h_17  -13.60 -12.91 -12.61 -12.30 -11.628
#> h_18  -13.69 -12.98 -12.65 -12.34 -11.667
#> h_19  -13.71 -13.01 -12.68 -12.37 -11.771
#> h_20  -13.63 -13.00 -12.68 -12.37 -11.678
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean        SD  Naive SE Time-series SE
#> h_1   0.002228 0.0007221 1.615e-05      2.997e-05
#> h_2   0.002247 0.0006729 1.505e-05      2.895e-05
#> h_3   0.002278 0.0006430 1.438e-05      2.845e-05
#> h_4   0.002284 0.0006501 1.454e-05      3.029e-05
#> h_5   0.002317 0.0006216 1.390e-05      2.720e-05
#> h_6   0.002403 0.0005990 1.339e-05      2.654e-05
#> h_7   0.002524 0.0005690 1.272e-05      2.510e-05
#> h_8   0.002599 0.0005682 1.271e-05      2.769e-05
#> h_9   0.002597 0.0005703 1.275e-05      2.432e-05
#> h_10  0.002638 0.0005615 1.256e-05      2.664e-05
#> h_11  0.002542 0.0005378 1.202e-05      2.320e-05
#> h_12  0.002330 0.0005421 1.212e-05      2.040e-05
#> h_13  0.002173 0.0005346 1.195e-05      2.168e-05
#> h_14  0.002054 0.0005175 1.157e-05      2.102e-05
#> h_15  0.001989 0.0005074 1.135e-05      2.278e-05
#> h_16  0.001955 0.0004928 1.102e-05      2.364e-05
#> h_17  0.001890 0.0004771 1.067e-05      2.069e-05
#> h_18  0.001838 0.0004688 1.048e-05      2.167e-05
#> h_19  0.001802 0.0004487 1.003e-05      1.940e-05
#> h_20  0.001815 0.0004588 1.026e-05      1.993e-05
#> h_21  0.001816 0.0004370 9.772e-06      1.941e-05
#> h_22  0.001844 0.0004203 9.398e-06      1.994e-05
#> h_23  0.001860 0.0004200 9.391e-06      2.084e-05
#> h_24  0.001821 0.0004136 9.248e-06      1.939e-05
#> h_25  0.001747 0.0004198 9.387e-06      1.737e-05
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>            2.5%      25%      50%      75%    97.5%
#> h_1   0.0011212 0.001737 0.002122 0.002596 0.003920
#> h_2   0.0012433 0.001778 0.002139 0.002585 0.003847
#> h_3   0.0013129 0.001815 0.002180 0.002610 0.003804
#> h_4   0.0013003 0.001829 0.002200 0.002628 0.003779
#> h_5   0.0013234 0.001884 0.002246 0.002639 0.003788
#> h_6   0.0014419 0.001984 0.002317 0.002735 0.003822
#> h_7   0.0016754 0.002113 0.002448 0.002822 0.003880
#> h_8   0.0017148 0.002201 0.002510 0.002916 0.003963
#> h_9   0.0016990 0.002195 0.002533 0.002906 0.003974
#> h_10  0.0017761 0.002232 0.002564 0.002944 0.003963
#> h_11  0.0016714 0.002167 0.002477 0.002827 0.003759
#> h_12  0.0014517 0.001965 0.002280 0.002622 0.003607
#> h_13  0.0012956 0.001804 0.002110 0.002471 0.003347
#> h_14  0.0011863 0.001707 0.001991 0.002352 0.003214
#> h_15  0.0011231 0.001646 0.001938 0.002256 0.003141
#> h_16  0.0011754 0.001611 0.001893 0.002208 0.003100
#> h_17  0.0011156 0.001572 0.001828 0.002133 0.002985
#> h_18  0.0010663 0.001516 0.001793 0.002091 0.002927
#> h_19  0.0010548 0.001496 0.001762 0.002065 0.002779
#> h_20  0.0010972 0.001501 0.001762 0.002064 0.002912
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,] -1.406083e-04 -1.542297e-01 -1.260157e-01
#>     [2,] -4.059357e-06  5.004300e-02 -5.314836e-02
#>     [3,]  1.536219e-04  2.217420e-01 -1.552207e-01
#>     [4,] -7.245712e-05  5.061933e-02  3.776379e-02
#>     [5,] -1.091780e-04  7.172346e-02  4.845293e-02
#>     [6,] -1.105059e-04  9.450898e-02 -1.286118e-03
#>     [7,] -4.212095e-05  1.807287e-01 -1.471112e-01
#>     [8,] -1.699213e-04  1.065601e-01  1.275752e-01
#>     [9,] -2.267998e-04  6.821944e-02 -8.553237e-02
#>    [10,] -2.179524e-04  8.908046e-03 -1.302423e-01
#>    [11,] -1.614190e-04  9.786260e-03 -8.005654e-02
#>    [12,] -4.160765e-05  6.134178e-02  2.313807e-02
#>    [13,]  1.833786e-05  1.352741e-01 -1.387181e-01
#>    [14,] -2.926052e-04 -3.617899e-02 -6.693810e-02
#>    [15,] -2.422444e-04 -1.398804e-02 -1.033773e-01
#>    [16,] -2.049423e-05  2.092239e-02 -6.717392e-02
#>    [17,] -8.350314e-05  3.674344e-02  3.365457e-02
#>    [18,] -2.222206e-04 -1.577727e-02 -5.491358e-03
#>    [19,] -2.293763e-04  2.999962e-03 -1.890639e-02
#>    [20,]  5.051560e-05  2.428760e-02 -1.164475e-01
#>    [21,] -3.489082e-04  6.043088e-02  3.584656e-03
#>    [22,] -1.761860e-04  1.500837e-01 -1.074997e-01
#>    [23,]  4.814458e-05  1.497158e-01 -1.027345e-01
#>    [24,]  6.564672e-05  8.350281e-02 -1.026730e-01
#>    [25,] -1.320916e-04  2.378383e-02  1.011588e-01
#>    [26,] -1.927088e-04  6.287520e-02 -1.112128e-02
#>    [27,] -3.830416e-04 -2.388530e-02 -4.489658e-02
#>    [28,] -6.920323e-04  2.300810e-02 -2.184807e-02
#>    [29,] -3.051879e-04 -2.648106e-02 -1.492489e-02
#>    [30,] -3.827977e-04 -1.619106e-02 -8.054390e-02
#>    [31,] -3.413286e-04 -9.488502e-03 -4.897948e-02
#>    [32,] -7.165938e-05 -7.007303e-02  5.601050e-02
#>    [33,] -2.086788e-04  1.508828e-01  2.030493e-02
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>               tau_1     tau_2      tau_3     tau_4     tau_5     tau_6
#>               tau_7      tau_8      tau_9     tau_10      tau_11    tau_12
#>              tau_13     tau_14     tau_15     tau_16    tau_17     tau_18
#>             tau_19     tau_20     tau_21    tau_22     tau_23    tau_24
#>              tau_25     tau_26    tau_27    tau_28    tau_29     tau_30
#>              tau_31     tau_32     tau_33    tau_34     tau_35    tau_36
#>             tau_37    tau_38     tau_39     tau_40    tau_41    tau_42
#>             tau_43    tau_44     tau_45    tau_46     tau_47     tau_48
#>             tau_49    tau_50    tau_51    tau_52     tau_53     tau_54
#>             tau_55    tau_56    tau_57     tau_58     tau_59     tau_60
#>              tau_61     tau_62     tau_63     tau_64     tau_65     tau_66
#>             tau_67     tau_68    tau_69     tau_70     tau_71    tau_72
#>             tau_73      tau_74     tau_75    tau_76    tau_77     tau_78
#>              tau_79    tau_80     tau_81    tau_82      tau_83     tau_84
#>              tau_85     tau_86     tau_87    tau_88     tau_89    tau_90
#>              tau_91    tau_92     tau_93     tau_94     tau_95     tau_96
#>             tau_97     tau_98     tau_99    tau_100   tau_101   tau_102
#>            tau_103   tau_104    tau_105   tau_106    tau_107    tau_108
#>             tau_109    tau_110   tau_111   tau_112   tau_113    tau_114
#>             tau_115    tau_116   tau_117   tau_118    tau_119   tau_120
#>             tau_121    tau_122    tau_123   tau_124    tau_125    tau_126
#>            tau_127    tau_128    tau_129    tau_130    tau_131   tau_132
#>             tau_133    tau_134   tau_135    tau_136   tau_137    tau_138
#>            tau_139    tau_140    tau_141    tau_142    tau_143   tau_144
#>             tau_145   tau_146   tau_147    tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }