Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean      SD Naive SE Time-series SE
#> mu    -9.8028 0.56001 0.012522       0.018466
#> phi    0.8878 0.08134 0.001819       0.009013
#> sigma  0.3286 0.11338 0.002535       0.012891
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%     50%     75%   97.5%
#> mu    -10.4910 -10.0528 -9.8558 -9.6514 -8.8953
#> phi     0.6672   0.8574  0.9080  0.9415  0.9845
#> sigma   0.1535   0.2442  0.3127  0.3966  0.5907
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>       -9.48066        0.61483        0.01375        0.02388 
#> 
#> 2. Quantiles for each variable:
#> 
#>    2.5%     25%     50%     75%   97.5% 
#> -10.630  -9.892  -9.521  -9.107  -8.203 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> h_1    -9.432 0.5551 0.012413        0.02411
#> h_2    -9.459 0.5645 0.012623        0.02664
#> h_3    -9.445 0.5620 0.012568        0.02584
#> h_4    -9.369 0.5435 0.012153        0.02590
#> h_5    -9.245 0.4985 0.011146        0.02367
#> h_6    -9.086 0.4563 0.010203        0.02173
#> h_7    -9.105 0.4836 0.010814        0.02492
#> h_8    -9.031 0.4671 0.010445        0.02038
#> h_9    -9.076 0.4944 0.011056        0.02106
#> h_10   -9.068 0.4654 0.010407        0.02380
#> h_11   -9.143 0.4896 0.010949        0.02471
#> h_12   -9.170 0.4962 0.011094        0.02533
#> h_13   -9.118 0.4763 0.010652        0.02426
#> h_14   -9.211 0.5077 0.011353        0.02241
#> h_15   -9.221 0.4951 0.011070        0.02206
#> h_16   -9.219 0.5010 0.011202        0.02210
#> h_17   -9.245 0.4995 0.011170        0.02216
#> h_18   -9.285 0.5062 0.011318        0.02035
#> h_19   -9.264 0.4998 0.011176        0.02136
#> h_20   -9.162 0.4580 0.010241        0.01827
#> h_21   -9.157 0.4748 0.010617        0.01927
#> h_22   -9.146 0.4631 0.010355        0.01885
#> h_23   -9.105 0.4570 0.010218        0.02046
#> h_24   -9.081 0.4372 0.009775        0.02162
#> h_25   -9.214 0.4512 0.010089        0.01946
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>          2.5%     25%     50%     75%  97.5%
#> h_1   -10.448  -9.795  -9.484  -9.094 -8.205
#> h_2   -10.486  -9.829  -9.483  -9.106 -8.273
#> h_3   -10.525  -9.799  -9.472  -9.082 -8.282
#> h_4   -10.416  -9.727  -9.375  -9.040 -8.243
#> h_5   -10.140  -9.578  -9.264  -8.929 -8.233
#> h_6    -9.919  -9.413  -9.116  -8.796 -8.110
#> h_7   -10.006  -9.425  -9.129  -8.797 -8.101
#> h_8    -9.860  -9.343  -9.067  -8.753 -8.032
#> h_9    -9.981  -9.407  -9.103  -8.750 -8.083
#> h_10   -9.914  -9.392  -9.081  -8.786 -8.081
#> h_11  -10.051  -9.470  -9.163  -8.821 -8.148
#> h_12  -10.125  -9.486  -9.186  -8.852 -8.178
#> h_13   -9.983  -9.441  -9.142  -8.823 -8.110
#> h_14  -10.206  -9.544  -9.215  -8.863 -8.227
#> h_15  -10.166  -9.562  -9.229  -8.903 -8.180
#> h_16  -10.140  -9.554  -9.229  -8.912 -8.180
#> h_17  -10.169  -9.578  -9.256  -8.943 -8.203
#> h_18  -10.247  -9.618  -9.297  -8.971 -8.249
#> h_19  -10.202  -9.583  -9.272  -8.957 -8.261
#> h_20  -10.001  -9.466  -9.179  -8.866 -8.172
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean       SD  Naive SE Time-series SE
#> h_1   0.009322 0.002893 6.468e-05      1.277e-04
#> h_2   0.009204 0.002828 6.324e-05      1.324e-04
#> h_3   0.009261 0.002771 6.196e-05      1.296e-04
#> h_4   0.009590 0.002758 6.167e-05      1.320e-04
#> h_5   0.010144 0.002671 5.973e-05      1.323e-04
#> h_6   0.010930 0.002642 5.909e-05      1.289e-04
#> h_7   0.010862 0.002814 6.292e-05      1.158e-04
#> h_8   0.011248 0.002816 6.297e-05      1.219e-04
#> h_9   0.011033 0.002855 6.384e-05      1.247e-04
#> h_10  0.011041 0.002698 6.032e-05      1.149e-04
#> h_11  0.010660 0.002712 6.065e-05      1.432e-04
#> h_12  0.010526 0.002709 6.058e-05      1.180e-04
#> h_13  0.010778 0.002698 6.033e-05      1.350e-04
#> h_14  0.010322 0.002680 5.992e-05      1.172e-04
#> h_15  0.010261 0.002681 5.994e-05      1.191e-04
#> h_16  0.010284 0.002727 6.098e-05      1.174e-04
#> h_17  0.010147 0.002697 6.031e-05      1.147e-04
#> h_18  0.009954 0.002693 6.021e-05      1.058e-04
#> h_19  0.010043 0.002593 5.798e-05      1.275e-04
#> h_20  0.010527 0.002553 5.710e-05      1.027e-04
#> h_21  0.010573 0.002687 6.008e-05      1.076e-04
#> h_22  0.010616 0.002593 5.797e-05      1.039e-04
#> h_23  0.010826 0.002621 5.861e-05      1.171e-04
#> h_24  0.010936 0.002568 5.741e-05      1.258e-04
#> h_25  0.010247 0.002497 5.584e-05      1.090e-04
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.005387 0.007466 0.008722 0.010596 0.016531
#> h_2   0.005285 0.007339 0.008727 0.010534 0.015980
#> h_3   0.005183 0.007449 0.008775 0.010665 0.015906
#> h_4   0.005471 0.007724 0.009209 0.010887 0.016218
#> h_5   0.006281 0.008319 0.009736 0.011512 0.016302
#> h_6   0.007016 0.009036 0.010484 0.012301 0.017333
#> h_7   0.006717 0.008984 0.010414 0.012297 0.017416
#> h_8   0.007227 0.009359 0.010744 0.012567 0.018026
#> h_9   0.006801 0.009066 0.010550 0.012587 0.017568
#> h_10  0.007034 0.009132 0.010670 0.012366 0.017591
#> h_11  0.006569 0.008780 0.010238 0.012148 0.017012
#> h_12  0.006330 0.008714 0.010124 0.011960 0.016755
#> h_13  0.006796 0.008911 0.010350 0.012135 0.017340
#> h_14  0.006078 0.008463 0.009978 0.011896 0.016347
#> h_15  0.006202 0.008389 0.009906 0.011661 0.016739
#> h_16  0.006283 0.008422 0.009909 0.011610 0.016736
#> h_17  0.006191 0.008320 0.009774 0.011428 0.016550
#> h_18  0.005954 0.008156 0.009574 0.011271 0.016173
#> h_19  0.006091 0.008302 0.009695 0.011351 0.016074
#> h_20  0.006735 0.008800 0.010156 0.011880 0.016807
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,] -1.014302e-03  1.286392e-01  4.861595e-02
#>     [2,] -9.573637e-04  8.624186e-03  2.601012e-01
#>     [3,] -8.270929e-04  5.751362e-02  1.354409e-01
#>     [4,] -1.246913e-04  1.287357e-01  1.703377e-01
#>     [5,] -7.592819e-04  1.031336e-01 -6.866616e-02
#>     [6,] -1.729542e-04 -5.266262e-02  6.874404e-02
#>     [7,] -1.225469e-04 -9.104762e-02  8.325832e-02
#>     [8,] -3.481612e-04 -1.195308e-01  2.321019e-01
#>     [9,]  3.282205e-06 -8.057467e-02  2.671039e-01
#>    [10,] -5.993647e-04  1.635430e-03  1.820320e-02
#>    [11,] -3.708260e-04 -1.015911e-01  1.343155e-01
#>    [12,] -4.247264e-04 -8.539949e-02  1.100274e-02
#>    [13,] -8.842272e-05 -2.992383e-02  6.760338e-02
#>    [14,] -2.596941e-04  7.167300e-02  2.816171e-02
#>    [15,] -5.072376e-04  3.213038e-02  2.323125e-01
#>    [16,] -8.816202e-05  8.062423e-02  2.152089e-01
#>    [17,]  4.520571e-04  1.307528e-01  2.356819e-01
#>    [18,] -5.023045e-04  3.365494e-02  5.020255e-02
#>    [19,]  7.079089e-05 -1.360383e-01  1.442736e-01
#>    [20,]  7.488173e-05 -2.013513e-02  6.283587e-02
#>    [21,] -1.002648e-04 -8.227827e-02  1.070412e-01
#>    [22,] -5.060080e-05 -3.779126e-02  2.288723e-01
#>    [23,] -7.882926e-05 -6.804851e-02  1.815391e-01
#>    [24,] -5.381803e-04 -1.019457e-01  2.743798e-01
#>    [25,] -7.633445e-04 -4.952331e-02  1.631134e-01
#>    [26,] -1.766630e-04 -9.744488e-03  2.150336e-01
#>    [27,]  4.095841e-04  9.965330e-02  9.629382e-06
#>    [28,]  3.353577e-04  4.142883e-02  1.662200e-01
#>    [29,] -4.603458e-04 -1.900840e-01  2.286599e-01
#>    [30,] -4.799250e-06  4.816378e-03  1.389281e-01
#>    [31,] -1.528277e-04  1.664318e-02  9.795876e-02
#>    [32,]  9.195420e-06  9.839057e-02  9.131161e-02
#>    [33,]  5.185834e-04  1.049324e-01  1.904491e-01
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>              tau_1     tau_2     tau_3      tau_4     tau_5      tau_6
#>               tau_7      tau_8     tau_9    tau_10     tau_11     tau_12
#>              tau_13     tau_14    tau_15    tau_16    tau_17     tau_18
#>              tau_19     tau_20     tau_21     tau_22    tau_23     tau_24
#>              tau_25     tau_26     tau_27    tau_28     tau_29    tau_30
#>              tau_31    tau_32     tau_33     tau_34    tau_35     tau_36
#>             tau_37    tau_38     tau_39    tau_40     tau_41    tau_42
#>              tau_43    tau_44    tau_45    tau_46    tau_47    tau_48    tau_49
#>              tau_50     tau_51     tau_52    tau_53     tau_54    tau_55
#>              tau_56     tau_57     tau_58     tau_59     tau_60     tau_61
#>              tau_62     tau_63    tau_64    tau_65     tau_66     tau_67
#>             tau_68    tau_69     tau_70    tau_71    tau_72    tau_73
#>              tau_74     tau_75    tau_76     tau_77     tau_78    tau_79
#>              tau_80     tau_81     tau_82    tau_83    tau_84      tau_85
#>             tau_86    tau_87    tau_88    tau_89    tau_90    tau_91     tau_92
#>              tau_93     tau_94    tau_95    tau_96    tau_97     tau_98
#>              tau_99   tau_100    tau_101    tau_102   tau_103   tau_104
#>             tau_105    tau_106   tau_107    tau_108   tau_109    tau_110
#>             tau_111   tau_112    tau_113   tau_114   tau_115   tau_116
#>             tau_117    tau_118    tau_119    tau_120    tau_121   tau_122
#>             tau_123   tau_124    tau_125    tau_126   tau_127   tau_128
#>             tau_129   tau_130    tau_131    tau_132   tau_133   tau_134
#>            tau_135    tau_136    tau_137   tau_138    tau_139   tau_140
#>             tau_141    tau_142    tau_143    tau_144   tau_145   tau_146
#>            tau_147    tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }