Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean      SD Naive SE Time-series SE
#> mu    -10.7539 0.43316 0.009686       0.018116
#> phi     0.8842 0.07375 0.001649       0.008455
#> sigma   0.4131 0.14823 0.003315       0.021211
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%   97.5%
#> mu    -11.6678 -10.9830 -10.7500 -10.5219 -9.8507
#> phi     0.7041   0.8447   0.9009   0.9381  0.9868
#> sigma   0.1771   0.2980   0.3929   0.5125  0.7247
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>      -10.33833        0.76795        0.01717        0.03337 
#> 
#> 2. Quantiles for each variable:
#> 
#>    2.5%     25%     50%     75%   97.5% 
#> -11.901 -10.842 -10.322  -9.842  -8.819 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> h_1   -10.292 0.6947  0.01553        0.03466
#> h_2   -10.232 0.6648  0.01486        0.03496
#> h_3   -10.050 0.6088  0.01361        0.02737
#> h_4    -9.820 0.5348  0.01196        0.02223
#> h_5    -9.569 0.4786  0.01070        0.02243
#> h_6    -9.679 0.5512  0.01233        0.02549
#> h_7    -9.688 0.5625  0.01258        0.02987
#> h_8    -9.650 0.5729  0.01281        0.03143
#> h_9    -9.532 0.5428  0.01214        0.02532
#> h_10   -9.398 0.5071  0.01134        0.02192
#> h_11   -9.396 0.5328  0.01191        0.01931
#> h_12   -9.312 0.5145  0.01150        0.02305
#> h_13   -9.283 0.5249  0.01174        0.02453
#> h_14   -9.260 0.5447  0.01218        0.02380
#> h_15   -9.243 0.5161  0.01154        0.02406
#> h_16   -9.325 0.5311  0.01188        0.02442
#> h_17   -9.379 0.5460  0.01221        0.02645
#> h_18   -9.409 0.5952  0.01331        0.02705
#> h_19   -9.338 0.5482  0.01226        0.02296
#> h_20   -9.332 0.5542  0.01239        0.02040
#> h_21   -9.230 0.5154  0.01152        0.02131
#> h_22   -9.086 0.4669  0.01044        0.02474
#> h_23   -9.389 0.5205  0.01164        0.02122
#> h_24   -9.640 0.5437  0.01216        0.02353
#> h_25   -9.833 0.5615  0.01256        0.02447
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>          2.5%     25%     50%     75%   97.5%
#> h_1   -11.681 -10.723 -10.303  -9.849  -8.905
#> h_2   -11.612 -10.640 -10.213  -9.798  -8.916
#> h_3   -11.330 -10.429 -10.047  -9.652  -8.857
#> h_4   -10.799 -10.177  -9.833  -9.497  -8.738
#> h_5   -10.404  -9.903  -9.612  -9.280  -8.498
#> h_6   -10.771 -10.030  -9.681  -9.327  -8.543
#> h_7   -10.848 -10.021  -9.688  -9.336  -8.559
#> h_8   -10.801 -10.011  -9.659  -9.285  -8.497
#> h_9   -10.591  -9.889  -9.538  -9.208  -8.394
#> h_10  -10.326  -9.744  -9.429  -9.084  -8.311
#> h_11  -10.347  -9.752  -9.429  -9.083  -8.289
#> h_12  -10.264  -9.652  -9.345  -8.999  -8.201
#> h_13  -10.232  -9.633  -9.325  -8.967  -8.196
#> h_14  -10.233  -9.636  -9.291  -8.925  -8.151
#> h_15  -10.144  -9.597  -9.268  -8.915  -8.175
#> h_16  -10.285  -9.689  -9.350  -8.981  -8.239
#> h_17  -10.412  -9.752  -9.394  -9.026  -8.315
#> h_18  -10.566  -9.787  -9.425  -9.027  -8.199
#> h_19  -10.332  -9.709  -9.353  -9.012  -8.204
#> h_20  -10.365  -9.699  -9.362  -8.970  -8.215
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean        SD  Naive SE Time-series SE
#> h_1   0.006188 0.0022805 5.099e-05      1.028e-04
#> h_2   0.006342 0.0021854 4.887e-05      9.647e-05
#> h_3   0.006884 0.0021785 4.871e-05      9.805e-05
#> h_4   0.007649 0.0021951 4.908e-05      9.244e-05
#> h_5   0.008612 0.0022354 4.998e-05      1.044e-04
#> h_6   0.008220 0.0023683 5.296e-05      1.045e-04
#> h_7   0.008190 0.0023328 5.216e-05      1.237e-04
#> h_8   0.008364 0.0025226 5.641e-05      1.270e-04
#> h_9   0.008841 0.0025424 5.685e-05      1.160e-04
#> h_10  0.009416 0.0025797 5.768e-05      1.114e-04
#> h_11  0.009458 0.0027457 6.140e-05      9.981e-05
#> h_12  0.009836 0.0027025 6.043e-05      1.218e-04
#> h_13  0.009994 0.0028541 6.382e-05      1.335e-04
#> h_14  0.010141 0.0030340 6.784e-05      1.431e-04
#> h_15  0.010186 0.0028377 6.345e-05      1.302e-04
#> h_16  0.009790 0.0027568 6.164e-05      1.256e-04
#> h_17  0.009546 0.0027097 6.059e-05      1.340e-04
#> h_18  0.009463 0.0029122 6.512e-05      1.297e-04
#> h_19  0.009758 0.0029428 6.580e-05      1.168e-04
#> h_20  0.009783 0.0028605 6.396e-05      9.986e-05
#> h_21  0.010247 0.0028365 6.343e-05      1.148e-04
#> h_22  0.010942 0.0027062 6.051e-05      1.409e-04
#> h_23  0.009468 0.0026065 5.828e-05      1.058e-04
#> h_24  0.008375 0.0023777 5.317e-05      1.025e-04
#> h_25  0.007621 0.0022123 4.947e-05      9.209e-05
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.002908 0.004693 0.005790 0.007266 0.011652
#> h_2   0.003010 0.004894 0.006056 0.007453 0.011588
#> h_3   0.003466 0.005438 0.006581 0.008017 0.011932
#> h_4   0.004520 0.006167 0.007324 0.008666 0.012665
#> h_5   0.005505 0.007072 0.008181 0.009660 0.014282
#> h_6   0.004583 0.006638 0.007903 0.009432 0.013959
#> h_7   0.004410 0.006669 0.007877 0.009392 0.013849
#> h_8   0.004515 0.006701 0.007991 0.009635 0.014282
#> h_9   0.005014 0.007123 0.008490 0.010010 0.015042
#> h_10  0.005726 0.007659 0.008965 0.010654 0.015675
#> h_11  0.005666 0.007628 0.008965 0.010656 0.015851
#> h_12  0.005906 0.008019 0.009348 0.011116 0.016563
#> h_13  0.006000 0.008096 0.009442 0.011296 0.016608
#> h_14  0.005996 0.008085 0.009603 0.011536 0.016980
#> h_15  0.006270 0.008243 0.009717 0.011590 0.016785
#> h_16  0.005844 0.007873 0.009328 0.011213 0.016256
#> h_17  0.005483 0.007628 0.009125 0.010966 0.015643
#> h_18  0.005077 0.007497 0.008984 0.010960 0.016581
#> h_19  0.005706 0.007793 0.009310 0.011045 0.016538
#> h_20  0.005614 0.007834 0.009270 0.011276 0.016447
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,]  6.709119e-05  1.944689e-02 -3.351888e-02
#>     [2,]  4.065746e-04 -1.494429e-01 -6.102925e-02
#>     [3,] -1.933502e-04 -1.575269e-01 -7.029425e-02
#>     [4,] -3.010084e-05 -3.104783e-01 -2.038701e-01
#>     [5,]  3.032029e-04 -4.708110e-02 -1.196254e-01
#>     [6,]  3.332175e-04 -1.202046e-01 -1.279718e-01
#>     [7,] -2.116904e-04 -1.692886e-01 -6.848810e-02
#>     [8,]  9.244885e-04 -9.052337e-02 -5.188979e-02
#>     [9,] -3.667424e-05  3.399727e-02 -1.767064e-01
#>    [10,]  1.976061e-04 -1.548694e-01  2.391040e-02
#>    [11,]  1.578909e-04 -6.099124e-02 -1.422295e-02
#>    [12,] -1.558568e-04 -6.827989e-02 -1.053398e-01
#>    [13,] -6.744068e-04 -1.135765e-01 -4.102174e-02
#>    [14,] -4.117490e-04 -2.517847e-02 -1.344756e-01
#>    [15,]  4.359913e-05 -6.977469e-02 -1.090710e-01
#>    [16,] -2.018006e-06  4.033854e-02  4.223902e-02
#>    [17,] -1.291391e-04  1.029601e-02 -2.019733e-02
#>    [18,] -5.591217e-05  7.201756e-02 -8.346456e-02
#>    [19,]  4.112170e-04  3.454609e-02 -1.581592e-01
#>    [20,]  3.439975e-04 -2.549264e-02 -1.401858e-01
#>    [21,] -1.542183e-04 -2.545137e-02 -3.906542e-02
#>    [22,]  8.265231e-05  8.852929e-02  2.876034e-02
#>    [23,] -3.138340e-04  7.104876e-02 -1.632708e-01
#>    [24,] -6.394972e-04 -1.387659e-02 -2.524529e-01
#>    [25,] -5.577664e-05 -5.955524e-02 -1.125532e-02
#>    [26,] -2.961535e-04 -9.759427e-02 -7.002350e-02
#>    [27,] -1.667880e-04  7.698407e-02 -1.150615e-01
#>    [28,] -4.129091e-04 -2.556176e-02 -1.563398e-01
#>    [29,] -4.498461e-04 -7.389774e-02 -1.579953e-01
#>    [30,] -3.401136e-04 -2.882372e-02 -1.865553e-01
#>    [31,] -1.511928e-04 -1.294894e-02 -2.076719e-01
#>    [32,] -5.603930e-04 -8.584581e-02 -5.096470e-02
#>    [33,] -1.363974e-04 -2.794226e-02 -3.118923e-02
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>              tau_1      tau_2      tau_3       tau_4     tau_5      tau_6
#>               tau_7      tau_8     tau_9     tau_10    tau_11     tau_12
#>              tau_13     tau_14      tau_15     tau_16     tau_17    tau_18
#>              tau_19     tau_20     tau_21    tau_22     tau_23     tau_24
#>              tau_25     tau_26     tau_27     tau_28     tau_29      tau_30
#>              tau_31     tau_32     tau_33     tau_34     tau_35     tau_36
#>             tau_37    tau_38     tau_39     tau_40     tau_41     tau_42
#>               tau_43     tau_44     tau_45    tau_46     tau_47     tau_48
#>             tau_49     tau_50    tau_51     tau_52     tau_53    tau_54
#>             tau_55     tau_56     tau_57    tau_58     tau_59     tau_60
#>             tau_61     tau_62     tau_63     tau_64     tau_65     tau_66
#>              tau_67     tau_68     tau_69     tau_70     tau_71    tau_72
#>              tau_73     tau_74     tau_75     tau_76     tau_77     tau_78
#>              tau_79     tau_80    tau_81    tau_82     tau_83     tau_84
#>              tau_85     tau_86      tau_87    tau_88    tau_89     tau_90
#>             tau_91     tau_92     tau_93     tau_94     tau_95     tau_96
#>              tau_97    tau_98    tau_99   tau_100   tau_101    tau_102
#>            tau_103    tau_104    tau_105    tau_106    tau_107    tau_108
#>             tau_109     tau_110    tau_111   tau_112    tau_113   tau_114
#>              tau_115   tau_116    tau_117   tau_118    tau_119    tau_120
#>             tau_121   tau_122     tau_123    tau_124    tau_125   tau_126
#>             tau_127    tau_128    tau_129    tau_130    tau_131    tau_132
#>             tau_133   tau_134    tau_135   tau_136    tau_137    tau_138
#>            tau_139   tau_140    tau_141    tau_142    tau_143    tau_144
#>            tau_145    tau_146   tau_147   tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }