Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean      SD Naive SE Time-series SE
#> mu    -9.4729 0.47409 0.010601        0.01853
#> phi    0.9011 0.09024 0.002018        0.01065
#> sigma  0.3024 0.15059 0.003367        0.02302
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%     25%     50%     75%   97.5%
#> mu    -10.4392 -9.6930 -9.4532 -9.2530 -8.5205
#> phi     0.6484  0.8665  0.9295  0.9644  0.9922
#> sigma   0.1097  0.1902  0.2622  0.3835  0.6642
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>       -9.30892        0.61661        0.01379        0.02477 
#> 
#> 2. Quantiles for each variable:
#> 
#>    2.5%     25%     50%     75%   97.5% 
#> -10.543  -9.672  -9.294  -8.923  -8.099 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> h_1    -9.307 0.5910 0.013215        0.02458
#> h_2    -9.279 0.5470 0.012231        0.02372
#> h_3    -9.199 0.4860 0.010867        0.01838
#> h_4    -9.131 0.4650 0.010398        0.01960
#> h_5    -9.016 0.4188 0.009364        0.01739
#> h_6    -9.053 0.4214 0.009423        0.01552
#> h_7    -9.142 0.4731 0.010579        0.01869
#> h_8    -9.173 0.4826 0.010790        0.02033
#> h_9    -9.184 0.5098 0.011400        0.02235
#> h_10   -9.129 0.5000 0.011181        0.02238
#> h_11   -9.037 0.4733 0.010582        0.02017
#> h_12   -8.924 0.4428 0.009901        0.01934
#> h_13   -8.826 0.4276 0.009563        0.02014
#> h_14   -8.889 0.4476 0.010009        0.01866
#> h_15   -8.865 0.4223 0.009442        0.02166
#> h_16   -8.962 0.4238 0.009476        0.01834
#> h_17   -9.036 0.4278 0.009566        0.01798
#> h_18   -9.204 0.5029 0.011245        0.02470
#> h_19   -9.298 0.5292 0.011834        0.02897
#> h_20   -9.328 0.5223 0.011678        0.03023
#> h_21   -9.334 0.5089 0.011380        0.03017
#> h_22   -9.370 0.5356 0.011976        0.03162
#> h_23   -9.347 0.5197 0.011621        0.02473
#> h_24   -9.308 0.4975 0.011125        0.02220
#> h_25   -9.269 0.5210 0.011649        0.02462
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>          2.5%     25%     50%     75%  97.5%
#> h_1   -10.525  -9.659  -9.270  -8.951 -8.179
#> h_2   -10.433  -9.613  -9.258  -8.928 -8.258
#> h_3   -10.143  -9.519  -9.202  -8.886 -8.237
#> h_4   -10.059  -9.418  -9.140  -8.855 -8.177
#> h_5    -9.781  -9.297  -9.044  -8.752 -8.143
#> h_6    -9.837  -9.333  -9.073  -8.785 -8.194
#> h_7   -10.151  -9.426  -9.138  -8.843 -8.227
#> h_8   -10.176  -9.470  -9.151  -8.876 -8.212
#> h_9   -10.338  -9.469  -9.158  -8.865 -8.223
#> h_10  -10.170  -9.421  -9.121  -8.830 -8.103
#> h_11   -9.916  -9.333  -9.045  -8.784 -8.011
#> h_12   -9.721  -9.202  -8.952  -8.677 -7.947
#> h_13   -9.579  -9.105  -8.865  -8.583 -7.870
#> h_14   -9.707  -9.181  -8.914  -8.629 -7.923
#> h_15   -9.603  -9.164  -8.893  -8.617 -7.919
#> h_16   -9.739  -9.245  -8.976  -8.710 -8.073
#> h_17   -9.831  -9.323  -9.051  -8.774 -8.145
#> h_18  -10.220  -9.487  -9.204  -8.896 -8.247
#> h_19  -10.478  -9.590  -9.274  -8.973 -8.345
#> h_20  -10.501  -9.615  -9.297  -8.995 -8.346
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean       SD  Naive SE Time-series SE
#> h_1   0.009949 0.002994 6.695e-05      1.250e-04
#> h_2   0.010030 0.002823 6.311e-05      1.156e-04
#> h_3   0.010362 0.002608 5.832e-05      9.668e-05
#> h_4   0.010694 0.002603 5.820e-05      1.098e-04
#> h_5   0.011274 0.002489 5.565e-05      9.915e-05
#> h_6   0.011064 0.002445 5.467e-05      9.015e-05
#> h_7   0.010639 0.002536 5.671e-05      9.745e-05
#> h_8   0.010484 0.002523 5.642e-05      1.029e-04
#> h_9   0.010459 0.002600 5.814e-05      1.103e-04
#> h_10  0.010750 0.002827 6.321e-05      1.104e-04
#> h_11  0.011222 0.002850 6.372e-05      1.139e-04
#> h_12  0.011836 0.002866 6.409e-05      1.262e-04
#> h_13  0.012409 0.002874 6.426e-05      1.358e-04
#> h_14  0.012049 0.002884 6.449e-05      1.207e-04
#> h_15  0.012160 0.002759 6.170e-05      1.348e-04
#> h_16  0.011588 0.002596 5.804e-05      1.121e-04
#> h_17  0.011168 0.002533 5.665e-05      1.052e-04
#> h_18  0.010345 0.002588 5.786e-05      1.221e-04
#> h_19  0.009899 0.002574 5.755e-05      1.236e-04
#> h_20  0.009744 0.002470 5.524e-05      1.291e-04
#> h_21  0.009706 0.002465 5.512e-05      1.349e-04
#> h_22  0.009562 0.002531 5.660e-05      1.382e-04
#> h_23  0.009651 0.002468 5.519e-05      1.134e-04
#> h_24  0.009820 0.002474 5.532e-05      1.048e-04
#> h_25  0.010038 0.002587 5.784e-05      1.166e-04
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.005182 0.007990 0.009708 0.011387 0.016746
#> h_2   0.005427 0.008178 0.009763 0.011516 0.016101
#> h_3   0.006272 0.008570 0.010043 0.011759 0.016268
#> h_4   0.006544 0.009014 0.010357 0.011947 0.016766
#> h_5   0.007518 0.009575 0.010866 0.012575 0.017052
#> h_6   0.007309 0.009404 0.010712 0.012372 0.016624
#> h_7   0.006249 0.008976 0.010370 0.012014 0.016351
#> h_8   0.006171 0.008783 0.010300 0.011820 0.016473
#> h_9   0.005691 0.008787 0.010268 0.011887 0.016382
#> h_10  0.006187 0.009002 0.010459 0.012096 0.017394
#> h_11  0.007027 0.009405 0.010861 0.012373 0.018211
#> h_12  0.007748 0.010042 0.011377 0.013059 0.018808
#> h_13  0.008315 0.010538 0.011886 0.013685 0.019545
#> h_14  0.007799 0.010148 0.011600 0.013373 0.019038
#> h_15  0.008219 0.010232 0.011720 0.013452 0.019069
#> h_16  0.007679 0.009829 0.011246 0.012842 0.017659
#> h_17  0.007332 0.009451 0.010828 0.012439 0.017036
#> h_18  0.006037 0.008707 0.010031 0.011702 0.016190
#> h_19  0.005307 0.008270 0.009689 0.011259 0.015413
#> h_20  0.005244 0.008169 0.009576 0.011136 0.015403
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,]  1.367264e-04  6.713458e-02  2.564709e-01
#>     [2,]  1.355978e-04  6.382080e-02  2.757573e-01
#>     [3,]  1.253544e-03  2.597110e-01  1.200666e-01
#>     [4,]  1.442468e-03  2.285455e-01  1.096339e-01
#>     [5,]  5.663230e-04  2.784520e-01  9.953629e-02
#>     [6,]  1.142222e-03  1.355948e-01  9.825315e-02
#>     [7,]  8.363703e-04  2.158329e-02  3.435616e-01
#>     [8,]  2.828730e-04  5.162492e-03  1.801150e-01
#>     [9,]  1.331859e-03 -3.933652e-02  2.881522e-01
#>    [10,]  9.328280e-04  1.097073e-01  4.424821e-02
#>    [11,]  2.257745e-03 -2.101570e-02  1.540529e-01
#>    [12,]  1.357394e-05  1.510601e-01  2.626832e-01
#>    [13,]  1.442383e-03  1.378608e-01  2.228452e-01
#>    [14,]  1.307280e-03  2.264546e-01  2.998415e-01
#>    [15,]  6.664453e-04  1.568665e-01  1.500007e-01
#>    [16,]  1.300920e-03  1.299921e-01  1.543531e-01
#>    [17,]  1.923166e-03  9.610562e-02  1.680729e-01
#>    [18,]  7.145275e-04 -5.925538e-02  2.135352e-01
#>    [19,]  9.970639e-04  2.091636e-01  7.815070e-02
#>    [20,]  4.809379e-04  2.576295e-01  1.274891e-01
#>    [21,]  9.681574e-04  8.271496e-02  3.085327e-01
#>    [22,]  1.366909e-04  1.131136e-01  1.485845e-01
#>    [23,]  5.042798e-04  5.670192e-02  1.386025e-01
#>    [24,]  1.690535e-03 -5.058034e-03  2.962911e-02
#>    [25,]  1.910547e-04  1.329174e-01  2.506791e-01
#>    [26,] -5.378825e-04  1.500745e-01  1.008756e-01
#>    [27,] -4.573673e-04  7.385684e-02  1.729003e-01
#>    [28,] -2.650915e-04  7.316364e-02  2.410543e-01
#>    [29,]  1.118405e-03  4.307351e-02  1.324116e-01
#>    [30,] -5.629995e-04  1.506702e-01  1.332938e-01
#>    [31,] -1.289121e-03  3.696757e-02  2.645195e-01
#>    [32,] -2.246236e-04  1.040424e-01  1.663497e-01
#>    [33,] -6.622406e-04  1.210326e-01  1.070429e-01
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>              tau_1     tau_2      tau_3     tau_4     tau_5     tau_6     tau_7
#>              tau_8     tau_9    tau_10    tau_11    tau_12    tau_13    tau_14
#>              tau_15    tau_16    tau_17    tau_18     tau_19    tau_20
#>             tau_21    tau_22     tau_23    tau_24     tau_25    tau_26
#>             tau_27    tau_28    tau_29    tau_30    tau_31    tau_32    tau_33
#>             tau_34     tau_35    tau_36    tau_37    tau_38    tau_39    tau_40
#>             tau_41    tau_42    tau_43    tau_44    tau_45     tau_46    tau_47
#>             tau_48     tau_49    tau_50    tau_51    tau_52    tau_53    tau_54
#>             tau_55    tau_56    tau_57    tau_58    tau_59    tau_60    tau_61
#>             tau_62    tau_63    tau_64    tau_65    tau_66    tau_67    tau_68
#>             tau_69    tau_70    tau_71    tau_72    tau_73    tau_74    tau_75
#>             tau_76     tau_77    tau_78    tau_79    tau_80    tau_81    tau_82
#>             tau_83    tau_84     tau_85    tau_86    tau_87    tau_88    tau_89
#>             tau_90    tau_91     tau_92     tau_93    tau_94    tau_95
#>             tau_96    tau_97    tau_98    tau_99    tau_100   tau_101   tau_102
#>             tau_103   tau_104   tau_105    tau_106    tau_107    tau_108
#>            tau_109   tau_110   tau_111   tau_112   tau_113   tau_114   tau_115
#>             tau_116   tau_117   tau_118   tau_119   tau_120    tau_121
#>             tau_122   tau_123   tau_124   tau_125   tau_126   tau_127
#>             tau_128   tau_129   tau_130    tau_131   tau_132   tau_133
#>            tau_134   tau_135   tau_136   tau_137    tau_138   tau_139   tau_140
#>            tau_141   tau_142   tau_143   tau_144   tau_145   tau_146   tau_147
#>             tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }