Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean     SD Naive SE Time-series SE
#> mu    -10.4439 0.2554 0.005712        0.01117
#> phi     0.5810 0.2967 0.006634        0.05451
#> sigma   0.4452 0.1939 0.004335        0.02514
#> 
#> 2. Quantiles for each variable:
#> 
#>            2.5%      25%      50%      75%   97.5%
#> mu    -10.90073 -10.5809 -10.4536 -10.3216 -9.9665
#> phi    -0.07417   0.3839   0.6201   0.8383  0.9743
#> sigma   0.12398   0.2969   0.4333   0.5816  0.8396
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>      -10.52536        0.63310        0.01416        0.02014 
#> 
#> 2. Quantiles for each variable:
#> 
#>    2.5%     25%     50%     75%   97.5% 
#> -11.837 -10.893 -10.507 -10.110  -9.325 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> h_1   -10.588 0.6107  0.01365        0.02184
#> h_2   -10.489 0.5317  0.01189        0.01912
#> h_3   -10.400 0.5052  0.01130        0.01888
#> h_4   -10.239 0.4719  0.01055        0.01886
#> h_5   -10.487 0.5499  0.01230        0.01791
#> h_6   -10.468 0.5284  0.01182        0.01806
#> h_7   -10.486 0.5090  0.01138        0.01736
#> h_8   -10.572 0.5411  0.01210        0.02169
#> h_9   -10.540 0.5375  0.01202        0.01874
#> h_10  -10.500 0.5463  0.01222        0.01948
#> h_11  -10.304 0.4943  0.01105        0.01978
#> h_12  -10.409 0.5186  0.01160        0.02014
#> h_13  -10.659 0.5924  0.01325        0.02713
#> h_14  -10.701 0.5919  0.01323        0.02287
#> h_15  -10.741 0.6026  0.01348        0.03056
#> h_16  -10.693 0.5765  0.01289        0.02593
#> h_17  -10.685 0.5830  0.01304        0.02437
#> h_18  -10.581 0.5658  0.01265        0.01961
#> h_19  -10.270 0.4664  0.01043        0.02434
#> h_20  -10.468 0.5146  0.01151        0.02438
#> h_21  -10.552 0.5725  0.01280        0.02694
#> h_22  -10.577 0.5566  0.01245        0.02680
#> h_23  -10.482 0.5310  0.01187        0.02094
#> h_24  -10.455 0.5395  0.01206        0.02181
#> h_25  -10.363 0.5124  0.01146        0.02076
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>         2.5%     25%     50%     75%  97.5%
#> h_1   -11.90 -10.942 -10.550 -10.207 -9.448
#> h_2   -11.57 -10.821 -10.499 -10.138 -9.436
#> h_3   -11.38 -10.731 -10.398 -10.083 -9.345
#> h_4   -11.08 -10.561 -10.260  -9.968 -9.254
#> h_5   -11.67 -10.816 -10.479 -10.141 -9.451
#> h_6   -11.53 -10.785 -10.467 -10.149 -9.415
#> h_7   -11.48 -10.823 -10.499 -10.166 -9.453
#> h_8   -11.61 -10.905 -10.570 -10.223 -9.499
#> h_9   -11.63 -10.862 -10.534 -10.220 -9.449
#> h_10  -11.61 -10.838 -10.500 -10.171 -9.415
#> h_11  -11.17 -10.635 -10.338  -9.998 -9.221
#> h_12  -11.41 -10.746 -10.432 -10.092 -9.350
#> h_13  -11.94 -11.010 -10.636 -10.277 -9.511
#> h_14  -12.00 -11.056 -10.658 -10.323 -9.570
#> h_15  -12.01 -11.125 -10.709 -10.325 -9.621
#> h_16  -11.90 -11.036 -10.670 -10.326 -9.569
#> h_17  -11.94 -11.042 -10.640 -10.313 -9.620
#> h_18  -11.71 -10.921 -10.567 -10.221 -9.500
#> h_19  -11.15 -10.572 -10.294  -9.995 -9.248
#> h_20  -11.47 -10.787 -10.469 -10.155 -9.433
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean       SD  Naive SE Time-series SE
#> h_1   0.005253 0.001598 3.572e-05      5.065e-05
#> h_2   0.005465 0.001472 3.290e-05      4.948e-05
#> h_3   0.005698 0.001498 3.350e-05      5.393e-05
#> h_4   0.006154 0.001577 3.527e-05      7.126e-05
#> h_5   0.005483 0.001530 3.422e-05      4.867e-05
#> h_6   0.005524 0.001531 3.424e-05      5.037e-05
#> h_7   0.005461 0.001466 3.277e-05      4.971e-05
#> h_8   0.005250 0.001465 3.275e-05      5.685e-05
#> h_9   0.005335 0.001509 3.373e-05      5.072e-05
#> h_10  0.005450 0.001570 3.511e-05      5.465e-05
#> h_11  0.005973 0.001579 3.532e-05      6.241e-05
#> h_12  0.005683 0.001569 3.507e-05      6.280e-05
#> h_13  0.005059 0.001485 3.320e-05      6.041e-05
#> h_14  0.004952 0.001444 3.228e-05      4.844e-05
#> h_15  0.004860 0.001441 3.222e-05      6.670e-05
#> h_16  0.004965 0.001457 3.258e-05      5.953e-05
#> h_17  0.004989 0.001470 3.288e-05      5.984e-05
#> h_18  0.005242 0.001494 3.341e-05      4.954e-05
#> h_19  0.006055 0.001509 3.374e-05      7.803e-05
#> h_20  0.005512 0.001469 3.284e-05      6.777e-05
#> h_21  0.005321 0.001519 3.397e-05      6.902e-05
#> h_22  0.005247 0.001467 3.280e-05      6.302e-05
#> h_23  0.005489 0.001543 3.451e-05      5.862e-05
#> h_24  0.005572 0.001619 3.621e-05      6.198e-05
#> h_25  0.005812 0.001585 3.544e-05      6.420e-05
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.002608 0.004208 0.005118 0.006077 0.008881
#> h_2   0.003068 0.004469 0.005251 0.006288 0.008932
#> h_3   0.003374 0.004675 0.005523 0.006465 0.009350
#> h_4   0.003924 0.005090 0.005918 0.006845 0.009785
#> h_5   0.002927 0.004480 0.005302 0.006279 0.008867
#> h_6   0.003132 0.004551 0.005336 0.006254 0.009029
#> h_7   0.003208 0.004465 0.005249 0.006201 0.008858
#> h_8   0.003015 0.004285 0.005067 0.006027 0.008656
#> h_9   0.002984 0.004379 0.005159 0.006035 0.008874
#> h_10  0.003020 0.004433 0.005248 0.006187 0.009029
#> h_11  0.003759 0.004905 0.005689 0.006745 0.009945
#> h_12  0.003326 0.004640 0.005428 0.006434 0.009326
#> h_13  0.002552 0.004067 0.004902 0.005865 0.008604
#> h_14  0.002477 0.003975 0.004849 0.005732 0.008353
#> h_15  0.002469 0.003840 0.004728 0.005726 0.008144
#> h_16  0.002606 0.004014 0.004821 0.005724 0.008357
#> h_17  0.002558 0.004001 0.004892 0.005761 0.008149
#> h_18  0.002870 0.004251 0.005076 0.006034 0.008651
#> h_19  0.003800 0.005061 0.005817 0.006754 0.009812
#> h_20  0.003236 0.004547 0.005329 0.006234 0.008948
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,]  1.803749e-06  5.467812e-02 -1.296039e-01
#>     [2,]  3.788988e-04  3.720351e-02 -8.899984e-03
#>     [3,] -5.808864e-04  1.950652e-01 -2.488197e-02
#>     [4,] -9.074631e-06  1.348336e-01 -1.143039e-02
#>     [5,]  5.304512e-04  4.526881e-02 -1.223119e-01
#>     [6,]  3.730994e-04  4.207682e-02 -2.483482e-02
#>     [7,] -4.876076e-04 -5.699200e-02  4.501206e-02
#>     [8,] -3.805266e-04 -9.858560e-02 -4.312410e-03
#>     [9,] -8.123093e-05 -6.105818e-02 -8.549686e-02
#>    [10,]  1.091406e-04 -5.276537e-02 -9.373545e-02
#>    [11,]  5.947298e-04 -1.267149e-02 -1.113777e-02
#>    [12,] -7.719973e-04  1.555629e-01  1.636324e-02
#>    [13,] -3.622635e-04  1.596201e-01 -3.779388e-02
#>    [14,] -4.411632e-04  1.721540e-01 -4.530686e-03
#>    [15,] -3.379114e-05  7.982487e-02 -1.002195e-01
#>    [16,] -3.519716e-04  7.024194e-02 -1.604177e-01
#>    [17,]  3.393451e-04  6.111459e-02 -1.233949e-01
#>    [18,]  4.155358e-04  1.491448e-01  1.154514e-01
#>    [19,]  5.483642e-04  1.921392e-01  5.097939e-03
#>    [20,] -3.480347e-04  6.278002e-02 -7.536032e-02
#>    [21,]  2.829432e-05 -7.076351e-02 -6.578985e-02
#>    [22,]  5.056629e-05 -6.309158e-02 -5.621453e-02
#>    [23,] -1.120912e-04  5.866738e-03 -3.549331e-02
#>    [24,] -2.302426e-04  5.057979e-02  1.231528e-02
#>    [25,] -4.038879e-04  1.675884e-02 -1.297925e-01
#>    [26,] -4.632175e-04  2.794443e-02 -3.824610e-02
#>    [27,] -2.700761e-08  8.253591e-02  1.206103e-01
#>    [28,]  4.501706e-04  9.438102e-02  4.051182e-02
#>    [29,] -1.063958e-03 -4.942693e-02 -3.596139e-02
#>    [30,] -6.239805e-04 -3.305861e-02 -1.341298e-02
#>    [31,] -6.930748e-04  4.019688e-02 -3.545557e-02
#>    [32,] -6.291743e-04  2.677414e-02 -6.384649e-02
#>    [33,] -4.051256e-05  1.122706e-01 -3.976480e-02
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>               tau_1      tau_2      tau_3      tau_4      tau_5      tau_6
#>               tau_7      tau_8      tau_9     tau_10      tau_11    tau_12
#>             tau_13     tau_14     tau_15    tau_16     tau_17     tau_18
#>              tau_19     tau_20     tau_21     tau_22     tau_23     tau_24
#>              tau_25     tau_26     tau_27     tau_28     tau_29     tau_30
#>              tau_31     tau_32     tau_33     tau_34     tau_35     tau_36
#>              tau_37     tau_38     tau_39    tau_40    tau_41     tau_42
#>              tau_43      tau_44     tau_45     tau_46     tau_47     tau_48
#>              tau_49    tau_50     tau_51     tau_52     tau_53     tau_54
#>               tau_55     tau_56     tau_57     tau_58     tau_59     tau_60
#>              tau_61     tau_62     tau_63    tau_64     tau_65      tau_66
#>              tau_67    tau_68     tau_69    tau_70     tau_71      tau_72
#>              tau_73     tau_74     tau_75     tau_76     tau_77     tau_78
#>              tau_79     tau_80    tau_81    tau_82     tau_83     tau_84
#>              tau_85      tau_86      tau_87     tau_88     tau_89      tau_90
#>              tau_91    tau_92     tau_93     tau_94     tau_95     tau_96
#>             tau_97     tau_98     tau_99    tau_100    tau_101    tau_102
#>            tau_103     tau_104    tau_105    tau_106    tau_107    tau_108
#>            tau_109    tau_110    tau_111    tau_112    tau_113    tau_114
#>             tau_115   tau_116   tau_117    tau_118    tau_119    tau_120
#>              tau_121    tau_122    tau_123   tau_124    tau_125    tau_126
#>             tau_127    tau_128    tau_129    tau_130   tau_131    tau_132
#>             tau_133    tau_134    tau_135    tau_136    tau_137    tau_138
#>             tau_139    tau_140    tau_141    tau_142    tau_143    tau_144
#>             tau_145   tau_146    tau_147    tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }