Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean     SD Naive SE Time-series SE
#> mu    -10.1279 0.9527 0.021304       0.044980
#> phi     0.9420 0.0509 0.001138       0.006136
#> sigma   0.2898 0.1165 0.002606       0.017559
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%     75%   97.5%
#> mu    -11.5003 -10.5441 -10.2496 -9.8477 -7.8196
#> phi     0.7952   0.9247   0.9559  0.9778  0.9963
#> sigma   0.1129   0.2111   0.2658  0.3515  0.5841
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>      -10.36188        0.67299        0.01505        0.02706 
#> 
#> 2. Quantiles for each variable:
#> 
#>    2.5%     25%     50%     75%   97.5% 
#> -11.766 -10.793 -10.362  -9.916  -9.042 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> h_1   -10.357 0.6344 0.014186        0.02611
#> h_2   -10.325 0.5742 0.012840        0.02430
#> h_3   -10.231 0.5105 0.011414        0.02200
#> h_4   -10.198 0.4813 0.010762        0.02062
#> h_5   -10.207 0.5019 0.011223        0.02297
#> h_6   -10.207 0.4876 0.010903        0.02163
#> h_7   -10.153 0.4567 0.010212        0.02031
#> h_8   -10.122 0.4506 0.010076        0.02164
#> h_9   -10.062 0.4386 0.009807        0.02617
#> h_10  -10.128 0.4380 0.009795        0.02486
#> h_11  -10.138 0.4232 0.009464        0.02776
#> h_12  -10.341 0.4487 0.010032        0.01797
#> h_13  -10.494 0.4655 0.010409        0.01957
#> h_14  -10.667 0.4947 0.011062        0.02089
#> h_15  -10.772 0.4988 0.011153        0.02176
#> h_16  -10.855 0.4931 0.011025        0.02279
#> h_17  -10.949 0.4877 0.010905        0.02344
#> h_18  -11.098 0.5251 0.011741        0.02739
#> h_19  -11.213 0.5432 0.012147        0.02806
#> h_20  -11.302 0.5621 0.012570        0.02991
#> h_21  -11.393 0.5977 0.013365        0.03522
#> h_22  -11.491 0.6278 0.014038        0.03694
#> h_23  -11.561 0.6453 0.014429        0.03923
#> h_24  -11.582 0.6478 0.014485        0.05176
#> h_25  -11.548 0.5974 0.013358        0.04664
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>          2.5%     25%     50%     75%   97.5%
#> h_1   -11.663 -10.755 -10.349  -9.978  -9.029
#> h_2   -11.435 -10.692 -10.332  -9.963  -9.194
#> h_3   -11.137 -10.578 -10.247  -9.900  -9.201
#> h_4   -11.097 -10.524 -10.218  -9.909  -9.183
#> h_5   -11.119 -10.559 -10.223  -9.893  -9.153
#> h_6   -11.107 -10.534 -10.218  -9.906  -9.209
#> h_7   -10.971 -10.470 -10.187  -9.881  -9.154
#> h_8   -10.948 -10.415 -10.153  -9.852  -9.165
#> h_9   -10.829 -10.359 -10.097  -9.789  -9.125
#> h_10  -10.904 -10.426 -10.158  -9.860  -9.211
#> h_11  -10.876 -10.432 -10.169  -9.878  -9.227
#> h_12  -11.144 -10.629 -10.368 -10.081  -9.361
#> h_13  -11.375 -10.789 -10.515 -10.206  -9.545
#> h_14  -11.657 -10.984 -10.666 -10.357  -9.721
#> h_15  -11.753 -11.075 -10.771 -10.453  -9.800
#> h_16  -11.800 -11.169 -10.858 -10.537  -9.928
#> h_17  -11.912 -11.278 -10.937 -10.629  -9.994
#> h_18  -12.137 -11.430 -11.087 -10.751 -10.086
#> h_19  -12.297 -11.566 -11.189 -10.848 -10.252
#> h_20  -12.497 -11.664 -11.270 -10.918 -10.302
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean        SD  Naive SE Time-series SE
#> h_1   0.005930 0.0019856 4.440e-05      8.058e-05
#> h_2   0.005975 0.0018460 4.128e-05      7.810e-05
#> h_3   0.006210 0.0017249 3.857e-05      7.211e-05
#> h_4   0.006287 0.0016166 3.615e-05      6.905e-05
#> h_5   0.006276 0.0016684 3.731e-05      7.600e-05
#> h_6   0.006261 0.0015918 3.559e-05      7.190e-05
#> h_7   0.006414 0.0015747 3.521e-05      7.271e-05
#> h_8   0.006507 0.0015767 3.526e-05      7.508e-05
#> h_9   0.006699 0.0015887 3.552e-05      9.535e-05
#> h_10  0.006481 0.0015279 3.417e-05      9.177e-05
#> h_11  0.006438 0.0014741 3.296e-05      9.428e-05
#> h_12  0.005832 0.0014001 3.131e-05      6.078e-05
#> h_13  0.005412 0.0013438 3.005e-05      5.581e-05
#> h_14  0.004980 0.0012916 2.888e-05      5.358e-05
#> h_15  0.004727 0.0012280 2.746e-05      5.254e-05
#> h_16  0.004531 0.0011729 2.623e-05      5.350e-05
#> h_17  0.004319 0.0010966 2.452e-05      5.403e-05
#> h_18  0.004029 0.0010921 2.442e-05      5.548e-05
#> h_19  0.003810 0.0010332 2.310e-05      5.139e-05
#> h_20  0.003653 0.0010180 2.276e-05      5.213e-05
#> h_21  0.003506 0.0010168 2.274e-05      5.649e-05
#> h_22  0.003353 0.0010198 2.280e-05      5.918e-05
#> h_23  0.003246 0.0010144 2.268e-05      5.887e-05
#> h_24  0.003212 0.0009955 2.226e-05      5.869e-05
#> h_25  0.003245 0.0009486 2.121e-05      7.059e-05
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.002934 0.004619 0.005658 0.006811 0.010949
#> h_2   0.003288 0.004768 0.005707 0.006863 0.010083
#> h_3   0.003816 0.005047 0.005956 0.007082 0.010046
#> h_4   0.003893 0.005185 0.006041 0.007050 0.010138
#> h_5   0.003851 0.005094 0.006028 0.007107 0.010292
#> h_6   0.003874 0.005159 0.006042 0.007061 0.010007
#> h_7   0.004146 0.005328 0.006138 0.007153 0.010286
#> h_8   0.004195 0.005475 0.006243 0.007256 0.010229
#> h_9   0.004452 0.005631 0.006420 0.007486 0.010434
#> h_10  0.004288 0.005446 0.006225 0.007226 0.009999
#> h_11  0.004349 0.005428 0.006192 0.007163 0.009919
#> h_12  0.003802 0.004919 0.005605 0.006471 0.009275
#> h_13  0.003388 0.004541 0.005209 0.006079 0.008461
#> h_14  0.002942 0.004121 0.004830 0.005638 0.007747
#> h_15  0.002805 0.003936 0.004582 0.005372 0.007445
#> h_16  0.002739 0.003755 0.004388 0.005152 0.006985
#> h_17  0.002590 0.003556 0.004217 0.004921 0.006757
#> h_18  0.002315 0.003296 0.003914 0.004629 0.006453
#> h_19  0.002137 0.003079 0.003718 0.004409 0.005941
#> h_20  0.001933 0.002932 0.003571 0.004258 0.005793
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,] -6.081449e-04 -1.850074e-01  4.527933e-02
#>     [2,] -4.770077e-04  2.434536e-03 -3.335792e-02
#>     [3,]  3.793027e-04 -1.172446e-01  1.685529e-03
#>     [4,] -3.333822e-04 -1.124085e-01 -1.006472e-02
#>     [5,]  8.627407e-05  7.941902e-02  1.830643e-02
#>     [6,] -3.102225e-04 -1.624842e-01 -1.129108e-04
#>     [7,]  3.720392e-04 -7.905333e-02 -1.819370e-02
#>     [8,]  9.278585e-05  3.255898e-03 -1.631039e-02
#>     [9,]  1.750748e-04 -5.749363e-02 -2.864513e-02
#>    [10,] -7.060785e-04 -4.178708e-02  7.587832e-02
#>    [11,]  2.384463e-04 -1.013955e-01  1.914755e-02
#>    [12,]  3.327312e-05 -5.612942e-02 -6.919878e-02
#>    [13,] -4.444652e-04 -4.307347e-02 -1.264804e-01
#>    [14,] -9.425445e-04 -6.954025e-02 -4.322354e-02
#>    [15,] -8.036926e-04 -7.546157e-02 -8.597259e-04
#>    [16,] -7.419459e-04 -1.373444e-01 -1.552773e-01
#>    [17,] -7.167837e-04 -2.346253e-02  1.387452e-02
#>    [18,] -5.256349e-04 -4.397400e-02 -3.035235e-02
#>    [19,] -9.732621e-05  2.093113e-01 -4.223884e-02
#>    [20,] -3.137222e-04 -5.524977e-02 -2.064997e-02
#>    [21,]  2.383181e-05  6.237783e-02 -2.290494e-02
#>    [22,]  3.197978e-05 -7.697171e-02 -9.494368e-02
#>    [23,]  7.353985e-05  6.941318e-02  4.366961e-02
#>    [24,] -5.576961e-04  1.999081e-01 -6.502154e-02
#>    [25,] -3.857104e-04  2.144805e-02  3.006911e-03
#>    [26,] -4.057321e-04  2.754956e-02 -8.249285e-02
#>    [27,] -5.249378e-04 -1.015219e-01 -8.473544e-03
#>    [28,] -3.746315e-04 -2.326993e-01 -1.300738e-01
#>    [29,] -1.053754e-03 -1.505739e-01 -2.138837e-01
#>    [30,] -8.292863e-04 -8.477793e-02 -1.812928e-02
#>    [31,] -5.569858e-04 -1.987285e-01 -2.188452e-02
#>    [32,] -5.907991e-04 -1.560524e-01 -1.841773e-01
#>    [33,] -1.312053e-03 -1.222752e-01 -3.126346e-01
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>               tau_1     tau_2     tau_3     tau_4     tau_5     tau_6
#>               tau_7     tau_8      tau_9    tau_10    tau_11     tau_12
#>             tau_13    tau_14    tau_15    tau_16     tau_17    tau_18    tau_19
#>             tau_20    tau_21     tau_22    tau_23    tau_24    tau_25
#>              tau_26    tau_27    tau_28     tau_29     tau_30     tau_31
#>              tau_32    tau_33    tau_34    tau_35    tau_36    tau_37    tau_38
#>             tau_39    tau_40    tau_41     tau_42    tau_43    tau_44
#>              tau_45    tau_46    tau_47    tau_48    tau_49    tau_50    tau_51
#>             tau_52    tau_53     tau_54    tau_55    tau_56    tau_57    tau_58
#>              tau_59    tau_60     tau_61     tau_62    tau_63    tau_64
#>             tau_65     tau_66    tau_67     tau_68     tau_69    tau_70
#>             tau_71    tau_72    tau_73     tau_74    tau_75    tau_76    tau_77
#>             tau_78    tau_79    tau_80    tau_81     tau_82    tau_83    tau_84
#>             tau_85    tau_86    tau_87    tau_88     tau_89    tau_90    tau_91
#>             tau_92    tau_93    tau_94     tau_95    tau_96     tau_97
#>             tau_98    tau_99   tau_100   tau_101   tau_102   tau_103   tau_104
#>            tau_105   tau_106   tau_107    tau_108    tau_109    tau_110
#>            tau_111   tau_112   tau_113   tau_114   tau_115    tau_116   tau_117
#>            tau_118    tau_119    tau_120    tau_121   tau_122   tau_123
#>             tau_124   tau_125   tau_126    tau_127   tau_128    tau_129
#>             tau_130    tau_131   tau_132   tau_133    tau_134   tau_135
#>             tau_136    tau_137   tau_138   tau_139   tau_140   tau_141
#>            tau_142    tau_143    tau_144   tau_145    tau_146   tau_147
#>             tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }