Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean     SD Naive SE Time-series SE
#> mu    -11.2687 0.1466 0.003278       0.007997
#> phi     0.5392 0.2392 0.005348       0.027582
#> sigma   0.2585 0.1503 0.003361       0.014534
#> 
#> 2. Quantiles for each variable:
#> 
#>             2.5%      25%      50%      75%    97.5%
#> mu    -11.574336 -11.3570 -11.2625 -11.1733 -11.0061
#> phi     0.007999   0.3927   0.5687   0.7181   0.9257
#> sigma   0.013201   0.1450   0.2504   0.3526   0.5886
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>     -11.253082       0.371008       0.008296       0.012405 
#> 
#> 2. Quantiles for each variable:
#> 
#>   2.5%    25%    50%    75%  97.5% 
#> -12.02 -11.45 -11.25 -11.05 -10.50 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>         Mean     SD Naive SE Time-series SE
#> h_1   -11.23 0.3509 0.007847        0.01168
#> h_2   -11.20 0.3417 0.007640        0.01233
#> h_3   -11.24 0.3645 0.008151        0.01253
#> h_4   -11.20 0.3345 0.007479        0.01336
#> h_5   -11.28 0.3408 0.007621        0.01329
#> h_6   -11.33 0.3672 0.008212        0.01349
#> h_7   -11.32 0.3656 0.008174        0.01334
#> h_8   -11.32 0.3690 0.008251        0.01405
#> h_9   -11.29 0.3554 0.007947        0.01357
#> h_10  -11.30 0.3685 0.008241        0.01514
#> h_11  -11.27 0.3544 0.007925        0.01448
#> h_12  -11.33 0.3733 0.008347        0.01549
#> h_13  -11.36 0.3962 0.008859        0.01904
#> h_14  -11.34 0.3678 0.008224        0.01485
#> h_15  -11.35 0.3715 0.008307        0.01648
#> h_16  -11.37 0.3860 0.008631        0.01603
#> h_17  -11.38 0.4047 0.009050        0.01840
#> h_18  -11.37 0.3714 0.008304        0.01506
#> h_19  -11.35 0.3846 0.008599        0.01541
#> h_20  -11.32 0.3665 0.008195        0.01404
#> h_21  -11.26 0.3449 0.007713        0.01258
#> h_22  -11.23 0.3510 0.007849        0.01485
#> h_23  -11.20 0.3522 0.007876        0.01329
#> h_24  -11.17 0.3290 0.007357        0.01459
#> h_25  -11.25 0.3527 0.007886        0.01246
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>         2.5%    25%    50%    75%   97.5%
#> h_1   -11.96 -11.43 -11.23 -11.03 -10.458
#> h_2   -11.86 -11.41 -11.21 -11.02 -10.486
#> h_3   -12.02 -11.44 -11.24 -11.05 -10.537
#> h_4   -11.85 -11.40 -11.21 -11.02 -10.452
#> h_5   -12.00 -11.47 -11.27 -11.09 -10.574
#> h_6   -12.15 -11.51 -11.29 -11.13 -10.617
#> h_7   -12.15 -11.53 -11.29 -11.10 -10.605
#> h_8   -12.14 -11.51 -11.30 -11.10 -10.600
#> h_9   -12.04 -11.49 -11.28 -11.08 -10.548
#> h_10  -12.09 -11.50 -11.28 -11.10 -10.531
#> h_11  -12.04 -11.47 -11.26 -11.06 -10.540
#> h_12  -12.19 -11.53 -11.30 -11.12 -10.644
#> h_13  -12.29 -11.56 -11.32 -11.13 -10.638
#> h_14  -12.12 -11.54 -11.31 -11.12 -10.619
#> h_15  -12.18 -11.54 -11.33 -11.14 -10.627
#> h_16  -12.30 -11.56 -11.33 -11.14 -10.648
#> h_17  -12.34 -11.57 -11.33 -11.14 -10.630
#> h_18  -12.23 -11.57 -11.32 -11.14 -10.684
#> h_19  -12.19 -11.55 -11.32 -11.12 -10.605
#> h_20  -12.14 -11.51 -11.29 -11.11 -10.626
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean        SD  Naive SE Time-series SE
#> h_1   0.003704 0.0006776 1.515e-05      2.228e-05
#> h_2   0.003748 0.0006679 1.493e-05      2.358e-05
#> h_3   0.003681 0.0006891 1.541e-05      2.477e-05
#> h_4   0.003755 0.0006559 1.467e-05      2.653e-05
#> h_5   0.003598 0.0006142 1.373e-05      2.445e-05
#> h_6   0.003529 0.0006494 1.452e-05      2.321e-05
#> h_7   0.003538 0.0006361 1.422e-05      2.229e-05
#> h_8   0.003535 0.0006552 1.465e-05      2.425e-05
#> h_9   0.003596 0.0006561 1.467e-05      2.495e-05
#> h_10  0.003581 0.0006678 1.493e-05      2.935e-05
#> h_11  0.003619 0.0006514 1.456e-05      2.601e-05
#> h_12  0.003518 0.0006575 1.470e-05      2.601e-05
#> h_13  0.003483 0.0006693 1.497e-05      3.007e-05
#> h_14  0.003507 0.0006394 1.430e-05      2.467e-05
#> h_15  0.003481 0.0006423 1.436e-05      2.483e-05
#> h_16  0.003457 0.0006476 1.448e-05      2.358e-05
#> h_17  0.003447 0.0006663 1.490e-05      2.818e-05
#> h_18  0.003455 0.0006298 1.408e-05      2.393e-05
#> h_19  0.003494 0.0006627 1.482e-05      2.472e-05
#> h_20  0.003543 0.0006388 1.428e-05      2.318e-05
#> h_21  0.003636 0.0006284 1.405e-05      2.246e-05
#> h_22  0.003705 0.0006749 1.509e-05      2.799e-05
#> h_23  0.003748 0.0006879 1.538e-05      2.630e-05
#> h_24  0.003811 0.0006605 1.477e-05      2.980e-05
#> h_25  0.003663 0.0006500 1.453e-05      2.335e-05
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.002534 0.003302 0.003641 0.004022 0.005358
#> h_2   0.002659 0.003333 0.003679 0.004045 0.005285
#> h_3   0.002450 0.003279 0.003622 0.003996 0.005150
#> h_4   0.002677 0.003353 0.003676 0.004046 0.005376
#> h_5   0.002479 0.003237 0.003562 0.003901 0.005058
#> h_6   0.002302 0.003166 0.003527 0.003837 0.004949
#> h_7   0.002304 0.003142 0.003529 0.003882 0.004979
#> h_8   0.002316 0.003163 0.003520 0.003878 0.004991
#> h_9   0.002434 0.003193 0.003561 0.003928 0.005123
#> h_10  0.002371 0.003178 0.003550 0.003895 0.005167
#> h_11  0.002434 0.003232 0.003583 0.003957 0.005143
#> h_12  0.002256 0.003137 0.003521 0.003855 0.004882
#> h_13  0.002148 0.003092 0.003487 0.003837 0.004897
#> h_14  0.002338 0.003116 0.003500 0.003850 0.004945
#> h_15  0.002264 0.003119 0.003471 0.003815 0.004923
#> h_16  0.002133 0.003089 0.003465 0.003819 0.004874
#> h_17  0.002090 0.003066 0.003465 0.003803 0.004917
#> h_18  0.002215 0.003079 0.003476 0.003803 0.004786
#> h_19  0.002253 0.003097 0.003478 0.003852 0.004980
#> h_20  0.002316 0.003168 0.003535 0.003874 0.004927
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,] -1.835370e-04  1.504371e-02  1.452719e-01
#>     [2,]  1.652267e-04  1.723440e-01  5.922129e-02
#>     [3,] -2.317663e-04 -2.430397e-03  6.458117e-02
#>     [4,] -3.852854e-04 -1.057798e-01  1.775165e-02
#>     [5,]  1.432458e-04 -9.969563e-02  9.836003e-02
#>     [6,]  4.195818e-06 -6.342986e-02  1.036770e-01
#>     [7,]  5.423356e-05  2.033098e-04  1.301125e-01
#>     [8,] -1.756828e-04  4.273190e-02  1.504980e-01
#>     [9,] -2.768200e-04 -7.390345e-03  1.642868e-01
#>    [10,]  5.869808e-05 -2.462459e-01  2.652060e-02
#>    [11,] -3.273870e-05  2.164104e-02  1.388273e-02
#>    [12,] -1.578868e-04  1.929465e-02  1.801406e-01
#>    [13,] -1.087187e-04 -4.869484e-03  3.637427e-02
#>    [14,]  5.100253e-04 -5.192143e-02  2.310204e-01
#>    [15,]  5.180920e-04 -3.811050e-02  2.316371e-01
#>    [16,] -9.300486e-06 -3.956487e-02  1.758591e-01
#>    [17,]  4.523172e-04  5.748450e-02  2.290930e-01
#>    [18,]  2.144834e-04 -1.084320e-01  1.516154e-01
#>    [19,]  1.782288e-04 -2.206922e-02 -4.123609e-02
#>    [20,] -1.846333e-05 -8.571804e-02  1.153120e-02
#>    [21,] -2.046231e-04  1.460727e-01 -7.805778e-03
#>    [22,]  4.424526e-04  5.617582e-02  1.007571e-01
#>    [23,] -3.196942e-04 -3.684878e-02  2.698103e-02
#>    [24,]  1.156359e-05  4.430338e-03  8.203958e-02
#>    [25,]  4.567029e-04  1.203657e-02  1.116101e-01
#>    [26,]  6.618034e-04  1.185679e-01  3.853857e-02
#>    [27,] -2.835419e-04  4.724823e-02  7.323500e-02
#>    [28,]  1.088477e-04  4.084700e-02  1.825147e-01
#>    [29,]  9.128242e-05  9.020679e-02  2.871930e-02
#>    [30,] -4.736246e-05  6.786694e-02  9.186045e-02
#>    [31,] -3.934513e-04  8.359141e-02  1.705536e-03
#>    [32,]  2.489865e-04 -1.088004e-01 -1.702634e-02
#>    [33,] -2.431839e-05 -9.517507e-02  2.476187e-01
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>              tau_1      tau_2     tau_3     tau_4     tau_5     tau_6
#>               tau_7     tau_8     tau_9    tau_10    tau_11     tau_12
#>             tau_13     tau_14    tau_15    tau_16    tau_17    tau_18
#>              tau_19    tau_20    tau_21     tau_22     tau_23    tau_24
#>             tau_25    tau_26    tau_27     tau_28    tau_29     tau_30
#>              tau_31    tau_32    tau_33    tau_34    tau_35     tau_36
#>             tau_37    tau_38     tau_39     tau_40     tau_41     tau_42
#>             tau_43    tau_44     tau_45     tau_46     tau_47     tau_48
#>              tau_49    tau_50    tau_51     tau_52    tau_53     tau_54
#>             tau_55    tau_56    tau_57     tau_58    tau_59     tau_60
#>             tau_61     tau_62     tau_63     tau_64     tau_65    tau_66
#>             tau_67    tau_68    tau_69     tau_70    tau_71    tau_72    tau_73
#>             tau_74    tau_75    tau_76    tau_77     tau_78     tau_79
#>             tau_80    tau_81    tau_82    tau_83     tau_84     tau_85
#>             tau_86    tau_87    tau_88     tau_89     tau_90     tau_91
#>             tau_92    tau_93    tau_94     tau_95     tau_96    tau_97
#>              tau_98    tau_99   tau_100   tau_101    tau_102    tau_103
#>             tau_104   tau_105   tau_106    tau_107    tau_108   tau_109
#>            tau_110   tau_111   tau_112   tau_113   tau_114   tau_115   tau_116
#>             tau_117    tau_118    tau_119   tau_120   tau_121   tau_122
#>             tau_123    tau_124    tau_125    tau_126    tau_127   tau_128
#>            tau_129    tau_130    tau_131    tau_132   tau_133    tau_134
#>            tau_135   tau_136   tau_137    tau_138    tau_139   tau_140
#>            tau_141   tau_142   tau_143    tau_144    tau_145   tau_146
#>            tau_147   tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }