Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean      SD Naive SE Time-series SE
#> mu    -10.8340 0.64934 0.014520       0.022341
#> phi     0.9182 0.06937 0.001551       0.008375
#> sigma   0.3557 0.13744 0.003073       0.021036
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%   97.5%
#> mu    -11.9533 -11.1338 -10.8461 -10.5614 -9.7056
#> phi     0.7353   0.8936   0.9369   0.9651  0.9928
#> sigma   0.1582   0.2552   0.3344   0.4305  0.7011
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>      -10.44510        0.78682        0.01759        0.04810 
#> 
#> 2. Quantiles for each variable:
#> 
#>    2.5%     25%     50%     75%   97.5% 
#> -12.029 -10.945 -10.429  -9.944  -8.939 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> h_1   -10.423 0.7516 0.016807        0.04136
#> h_2   -10.313 0.6715 0.015015        0.03679
#> h_3   -10.176 0.6059 0.013548        0.03159
#> h_4    -9.991 0.5143 0.011501        0.02401
#> h_5    -9.976 0.5156 0.011530        0.02434
#> h_6    -9.966 0.5413 0.012105        0.02828
#> h_7    -9.940 0.5755 0.012868        0.02900
#> h_8    -9.815 0.5465 0.012221        0.02851
#> h_9    -9.660 0.5158 0.011533        0.02572
#> h_10   -9.505 0.4998 0.011176        0.02544
#> h_11   -9.481 0.4904 0.010966        0.02577
#> h_12   -9.373 0.4874 0.010899        0.02511
#> h_13   -9.282 0.4980 0.011136        0.02966
#> h_14   -9.128 0.4721 0.010556        0.03020
#> h_15   -9.096 0.4605 0.010296        0.03290
#> h_16   -9.080 0.4355 0.009739        0.02803
#> h_17   -9.253 0.4567 0.010212        0.02706
#> h_18   -9.428 0.4975 0.011124        0.02838
#> h_19   -9.569 0.4982 0.011139        0.02709
#> h_20   -9.737 0.5458 0.012205        0.03076
#> h_21   -9.874 0.5690 0.012723        0.03287
#> h_22   -9.893 0.5403 0.012081        0.02770
#> h_23  -10.001 0.5608 0.012541        0.03188
#> h_24  -10.145 0.6329 0.014152        0.04323
#> h_25  -10.251 0.6899 0.015427        0.05230
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>          2.5%     25%     50%     75%   97.5%
#> h_1   -11.968 -10.886 -10.417  -9.928  -8.982
#> h_2   -11.678 -10.753 -10.323  -9.897  -8.952
#> h_3   -11.344 -10.570 -10.200  -9.790  -8.932
#> h_4   -10.884 -10.358 -10.043  -9.682  -8.863
#> h_5   -10.912 -10.331 -10.001  -9.659  -8.887
#> h_6   -11.017 -10.318  -9.975  -9.610  -8.878
#> h_7   -11.105 -10.293  -9.936  -9.580  -8.780
#> h_8   -10.838 -10.160  -9.819  -9.468  -8.686
#> h_9   -10.620 -10.014  -9.678  -9.343  -8.597
#> h_10  -10.375  -9.845  -9.525  -9.189  -8.431
#> h_11  -10.385  -9.812  -9.490  -9.177  -8.494
#> h_12  -10.285  -9.689  -9.390  -9.089  -8.331
#> h_13  -10.188  -9.619  -9.307  -8.992  -8.222
#> h_14   -9.980  -9.453  -9.156  -8.844  -8.088
#> h_15   -9.932  -9.403  -9.119  -8.815  -8.122
#> h_16   -9.856  -9.374  -9.117  -8.814  -8.119
#> h_17  -10.084  -9.568  -9.279  -8.966  -8.274
#> h_18  -10.367  -9.752  -9.441  -9.103  -8.401
#> h_19  -10.522  -9.901  -9.567  -9.251  -8.554
#> h_20  -10.880 -10.079  -9.723  -9.373  -8.691
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean        SD  Naive SE Time-series SE
#> h_1   0.005846 0.0022259 4.977e-05      1.259e-04
#> h_2   0.006095 0.0021106 4.719e-05      1.129e-04
#> h_3   0.006465 0.0020475 4.578e-05      1.052e-04
#> h_4   0.007003 0.0019368 4.331e-05      9.158e-05
#> h_5   0.007057 0.0019552 4.372e-05      9.663e-05
#> h_6   0.007113 0.0020242 4.526e-05      9.644e-05
#> h_7   0.007239 0.0021609 4.832e-05      1.020e-04
#> h_8   0.007675 0.0021956 4.909e-05      1.123e-04
#> h_9   0.008267 0.0023067 5.158e-05      1.161e-04
#> h_10  0.008914 0.0024271 5.427e-05      1.242e-04
#> h_11  0.009009 0.0023553 5.267e-05      1.235e-04
#> h_12  0.009505 0.0024898 5.567e-05      1.305e-04
#> h_13  0.009964 0.0027140 6.069e-05      1.632e-04
#> h_14  0.010725 0.0027113 6.063e-05      1.810e-04
#> h_15  0.010880 0.0026541 5.935e-05      1.708e-04
#> h_16  0.010938 0.0025812 5.772e-05      1.632e-04
#> h_17  0.010053 0.0024126 5.395e-05      1.323e-04
#> h_18  0.009254 0.0024208 5.413e-05      1.380e-04
#> h_19  0.008625 0.0022152 4.953e-05      1.203e-04
#> h_20  0.007971 0.0021682 4.848e-05      1.073e-04
#> h_21  0.007463 0.0020867 4.666e-05      1.130e-04
#> h_22  0.007373 0.0020583 4.603e-05      1.000e-04
#> h_23  0.007006 0.0020079 4.490e-05      1.082e-04
#> h_24  0.006577 0.0020282 4.535e-05      1.412e-04
#> h_25  0.006289 0.0020876 4.668e-05      1.477e-04
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.002519 0.004327 0.005470 0.006986 0.011208
#> h_2   0.002912 0.004625 0.005733 0.007094 0.011380
#> h_3   0.003441 0.005067 0.006098 0.007483 0.011491
#> h_4   0.004330 0.005635 0.006593 0.007898 0.011898
#> h_5   0.004271 0.005709 0.006735 0.007989 0.011756
#> h_6   0.004052 0.005747 0.006822 0.008188 0.011809
#> h_7   0.003877 0.005820 0.006957 0.008313 0.012402
#> h_8   0.004431 0.006218 0.007376 0.008792 0.012998
#> h_9   0.004941 0.006690 0.007915 0.009359 0.013586
#> h_10  0.005586 0.007279 0.008543 0.010106 0.014764
#> h_11  0.005558 0.007404 0.008693 0.010167 0.014304
#> h_12  0.005842 0.007872 0.009141 0.010626 0.015522
#> h_13  0.006132 0.008154 0.009530 0.011154 0.016395
#> h_14  0.006804 0.008857 0.010273 0.012009 0.017528
#> h_15  0.006972 0.009081 0.010468 0.012186 0.017234
#> h_16  0.007243 0.009212 0.010479 0.012191 0.017260
#> h_17  0.006462 0.008361 0.009660 0.011302 0.015969
#> h_18  0.005608 0.007626 0.008912 0.010551 0.014985
#> h_19  0.005191 0.007080 0.008367 0.009800 0.013882
#> h_20  0.004340 0.006476 0.007740 0.009221 0.012964
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,]  4.651105e-05 -1.822804e-01  1.083501e-01
#>     [2,] -2.130384e-04 -1.345260e-01  1.232800e-01
#>     [3,]  3.691910e-04  8.572306e-02  1.220533e-01
#>     [4,]  4.208843e-04 -1.109137e-02  2.966254e-01
#>     [5,]  2.643887e-04 -2.739916e-02  3.097343e-01
#>     [6,] -2.228767e-04 -1.104305e-01  3.116358e-02
#>     [7,] -4.687660e-04  3.657832e-02  5.680758e-02
#>     [8,]  5.914250e-05  1.376159e-02  1.214779e-01
#>     [9,]  2.074975e-04 -7.169999e-02  1.690483e-01
#>    [10,]  1.863484e-04 -4.760566e-02  2.386693e-01
#>    [11,] -2.446720e-05 -1.253925e-02  5.152790e-02
#>    [12,] -3.422173e-04  2.292374e-02  1.172862e-01
#>    [13,] -3.760188e-04 -5.287980e-02  8.698279e-02
#>    [14,] -4.530437e-04  3.829988e-02 -1.168756e-01
#>    [15,] -8.151114e-05  3.505079e-02  4.041500e-02
#>    [16,] -1.092765e-04  7.999459e-02 -1.764620e-03
#>    [17,] -1.623299e-04 -9.317702e-04  9.157863e-02
#>    [18,] -4.661266e-04 -1.282912e-01  1.371387e-01
#>    [19,]  2.745424e-04 -7.854243e-02  9.655786e-02
#>    [20,] -1.807066e-04 -6.716903e-02  1.278723e-01
#>    [21,] -7.995283e-04 -8.476204e-02  1.509462e-01
#>    [22,] -6.793067e-04  1.003089e-02  1.894450e-01
#>    [23,] -3.002071e-06 -4.437250e-02  6.077840e-02
#>    [24,] -3.993578e-04 -6.637184e-02  8.103475e-02
#>    [25,]  3.682541e-04 -6.674319e-02  2.214895e-01
#>    [26,]  6.747080e-04 -1.349189e-01  4.715533e-02
#>    [27,] -1.024618e-04 -1.113245e-01 -2.023115e-02
#>    [28,] -2.819051e-05 -1.151111e-01  1.102337e-01
#>    [29,] -1.621940e-04  1.792489e-02  7.820132e-02
#>    [30,] -6.186495e-05  1.681468e-01  7.622377e-02
#>    [31,] -2.858615e-04 -8.543658e-02  3.070498e-02
#>    [32,] -6.344650e-05 -3.050937e-03  8.131550e-02
#>    [33,]  1.036859e-04  1.011785e-01  1.437914e-01
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>              tau_1      tau_2      tau_3      tau_4     tau_5     tau_6
#>               tau_7      tau_8      tau_9     tau_10     tau_11     tau_12
#>              tau_13     tau_14     tau_15    tau_16     tau_17      tau_18
#>              tau_19     tau_20     tau_21     tau_22    tau_23      tau_24
#>              tau_25     tau_26     tau_27    tau_28     tau_29     tau_30
#>              tau_31     tau_32     tau_33     tau_34     tau_35    tau_36
#>              tau_37    tau_38     tau_39     tau_40     tau_41      tau_42
#>              tau_43     tau_44    tau_45     tau_46     tau_47    tau_48
#>              tau_49     tau_50     tau_51     tau_52    tau_53      tau_54
#>              tau_55    tau_56     tau_57     tau_58     tau_59      tau_60
#>              tau_61     tau_62    tau_63     tau_64     tau_65     tau_66
#>             tau_67     tau_68     tau_69     tau_70     tau_71    tau_72
#>              tau_73    tau_74    tau_75     tau_76     tau_77    tau_78
#>              tau_79     tau_80     tau_81    tau_82     tau_83     tau_84
#>              tau_85     tau_86     tau_87     tau_88     tau_89     tau_90
#>             tau_91    tau_92     tau_93     tau_94    tau_95     tau_96
#>              tau_97     tau_98     tau_99   tau_100    tau_101    tau_102
#>              tau_103    tau_104   tau_105    tau_106    tau_107   tau_108
#>             tau_109    tau_110   tau_111   tau_112    tau_113    tau_114
#>             tau_115    tau_116    tau_117   tau_118    tau_119    tau_120
#>             tau_121    tau_122    tau_123    tau_124     tau_125    tau_126
#>              tau_127    tau_128    tau_129    tau_130   tau_131   tau_132
#>             tau_133   tau_134   tau_135   tau_136    tau_137   tau_138
#>             tau_139    tau_140    tau_141    tau_142    tau_143    tau_144
#>             tau_145    tau_146    tau_147     tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }