Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean     SD Naive SE Time-series SE
#> mu    -10.5297 0.3069 0.006863        0.01254
#> phi     0.7964 0.1632 0.003650        0.02506
#> sigma   0.4484 0.1603 0.003586        0.02032
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%   97.5%
#> mu    -11.2178 -10.6906 -10.5095 -10.3431 -9.9953
#> phi     0.3029   0.7648   0.8419   0.8946  0.9633
#> sigma   0.2174   0.3375   0.4205   0.5194  0.8619
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>      -10.77562        0.74702        0.01670        0.02676 
#> 
#> 2. Quantiles for each variable:
#> 
#>    2.5%     25%     50%     75%   97.5% 
#> -12.318 -11.245 -10.764 -10.271  -9.306 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> h_1   -10.851 0.7021  0.01570        0.02922
#> h_2   -10.859 0.6861  0.01534        0.03454
#> h_3   -10.777 0.6160  0.01377        0.02839
#> h_4   -10.706 0.6365  0.01423        0.02865
#> h_5   -10.494 0.5663  0.01266        0.02517
#> h_6   -10.416 0.5466  0.01222        0.02383
#> h_7   -10.307 0.5266  0.01177        0.02264
#> h_8   -10.208 0.5340  0.01194        0.02586
#> h_9   -10.336 0.6197  0.01386        0.02812
#> h_10  -10.279 0.5979  0.01337        0.03367
#> h_11  -10.078 0.5458  0.01220        0.02697
#> h_12   -9.776 0.4748  0.01062        0.02585
#> h_13   -9.833 0.5194  0.01161        0.02148
#> h_14   -9.975 0.5673  0.01269        0.02745
#> h_15  -10.053 0.6480  0.01449        0.03691
#> h_16   -9.952 0.6105  0.01365        0.03076
#> h_17   -9.719 0.5091  0.01138        0.02112
#> h_18   -9.854 0.6238  0.01395        0.03101
#> h_19   -9.777 0.6224  0.01392        0.02840
#> h_20   -9.529 0.5363  0.01199        0.02191
#> h_21   -9.391 0.4780  0.01069        0.02178
#> h_22   -9.721 0.5695  0.01274        0.02571
#> h_23   -9.859 0.5469  0.01223        0.02351
#> h_24   -9.987 0.5514  0.01233        0.02331
#> h_25  -10.209 0.6062  0.01356        0.03178
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>          2.5%     25%     50%     75%   97.5%
#> h_1   -12.219 -11.323 -10.842 -10.374  -9.496
#> h_2   -12.232 -11.332 -10.828 -10.413  -9.502
#> h_3   -11.970 -11.197 -10.779 -10.367  -9.610
#> h_4   -11.937 -11.096 -10.701 -10.284  -9.464
#> h_5   -11.523 -10.886 -10.506 -10.135  -9.303
#> h_6   -11.451 -10.768 -10.410 -10.062  -9.320
#> h_7   -11.295 -10.659 -10.317  -9.967  -9.233
#> h_8   -11.194 -10.588 -10.236  -9.886  -9.071
#> h_9   -11.573 -10.708 -10.320  -9.965  -9.142
#> h_10  -11.531 -10.645 -10.256  -9.878  -9.150
#> h_11  -11.140 -10.429 -10.083  -9.724  -8.991
#> h_12  -10.629 -10.095  -9.809  -9.481  -8.762
#> h_13  -10.812 -10.171  -9.846  -9.521  -8.718
#> h_14  -11.109 -10.333  -9.993  -9.609  -8.852
#> h_15  -11.416 -10.462 -10.028  -9.637  -8.832
#> h_16  -11.262 -10.338  -9.948  -9.566  -8.745
#> h_17  -10.615 -10.057  -9.761  -9.402  -8.637
#> h_18  -11.225 -10.222  -9.858  -9.451  -8.600
#> h_19  -11.081 -10.145  -9.772  -9.375  -8.588
#> h_20  -10.508  -9.882  -9.563  -9.211  -8.411
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean       SD  Naive SE Time-series SE
#> h_1   0.004681 0.001684 3.765e-05      6.433e-05
#> h_2   0.004655 0.001691 3.781e-05      8.028e-05
#> h_3   0.004792 0.001532 3.427e-05      8.342e-05
#> h_4   0.004979 0.001632 3.650e-05      6.727e-05
#> h_5   0.005482 0.001640 3.666e-05      7.057e-05
#> h_6   0.005683 0.001614 3.609e-05      6.474e-05
#> h_7   0.005987 0.001654 3.699e-05      6.982e-05
#> h_8   0.006301 0.001838 4.110e-05      7.950e-05
#> h_9   0.005972 0.001887 4.219e-05      7.131e-05
#> h_10  0.006122 0.001809 4.045e-05      8.614e-05
#> h_11  0.006725 0.001883 4.211e-05      8.752e-05
#> h_12  0.007762 0.001986 4.441e-05      9.836e-05
#> h_13  0.007583 0.002112 4.723e-05      8.529e-05
#> h_14  0.007106 0.002099 4.693e-05      9.948e-05
#> h_15  0.006910 0.002285 5.108e-05      1.146e-04
#> h_16  0.007230 0.002271 5.078e-05      9.222e-05
#> h_17  0.008020 0.002243 5.015e-05      9.364e-05
#> h_18  0.007606 0.002401 5.369e-05      1.050e-04
#> h_19  0.007898 0.002458 5.496e-05      1.008e-04
#> h_20  0.008849 0.002548 5.698e-05      1.074e-04
#> h_21  0.009407 0.002366 5.291e-05      1.094e-04
#> h_22  0.008070 0.002441 5.458e-05      1.102e-04
#> h_23  0.007509 0.002146 4.799e-05      9.018e-05
#> h_24  0.007048 0.002013 4.501e-05      8.252e-05
#> h_25  0.006351 0.001939 4.336e-05      9.540e-05
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.002221 0.003477 0.004423 0.005588 0.008669
#> h_2   0.002207 0.003462 0.004454 0.005480 0.008642
#> h_3   0.002516 0.003704 0.004564 0.005608 0.008190
#> h_4   0.002558 0.003895 0.004746 0.005847 0.008809
#> h_5   0.003146 0.004326 0.005231 0.006299 0.009547
#> h_6   0.003262 0.004589 0.005489 0.006533 0.009468
#> h_7   0.003527 0.004847 0.005750 0.006850 0.009889
#> h_8   0.003709 0.005022 0.005989 0.007134 0.010720
#> h_9   0.003069 0.004729 0.005743 0.006858 0.010348
#> h_10  0.003134 0.004880 0.005929 0.007161 0.010304
#> h_11  0.003811 0.005436 0.006465 0.007737 0.011159
#> h_12  0.004921 0.006424 0.007413 0.008733 0.012513
#> h_13  0.004490 0.006186 0.007276 0.008562 0.012792
#> h_14  0.003870 0.005704 0.006762 0.008194 0.011959
#> h_15  0.003319 0.005349 0.006646 0.008080 0.012083
#> h_16  0.003585 0.005689 0.006915 0.008370 0.012623
#> h_17  0.004955 0.006550 0.007594 0.009087 0.013322
#> h_18  0.003651 0.006029 0.007234 0.008865 0.013570
#> h_19  0.003925 0.006266 0.007552 0.009211 0.013652
#> h_20  0.005227 0.007147 0.008382 0.009999 0.014912
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,] -8.585998e-04 -8.094750e-02 -6.551757e-03
#>     [2,] -7.717683e-04 -1.248536e-01 -3.112359e-02
#>     [3,] -9.976433e-04 -9.701240e-02 -4.600815e-02
#>     [4,] -1.295632e-03 -6.284049e-02  7.351528e-02
#>     [5,] -2.371334e-04 -1.144797e-02  1.396075e-01
#>     [6,]  1.372684e-04 -7.903572e-02  6.243000e-02
#>     [7,] -8.200006e-04  2.587243e-02  8.316710e-02
#>     [8,] -4.051198e-04 -1.027404e-01 -7.679311e-02
#>     [9,] -5.856523e-04 -1.894521e-01 -8.799733e-03
#>    [10,] -8.365460e-05 -1.303224e-02  1.821420e-02
#>    [11,] -5.225723e-04  7.620574e-02  9.412476e-02
#>    [12,] -3.463516e-04 -8.219093e-02  4.627592e-02
#>    [13,]  4.417109e-04 -5.366690e-02 -1.690931e-02
#>    [14,] -7.787960e-04  2.408987e-02 -1.038341e-01
#>    [15,] -3.968364e-04 -2.323001e-01  1.162366e-01
#>    [16,]  2.844209e-04 -6.899616e-02  2.709406e-02
#>    [17,] -3.332899e-04 -8.835869e-02  4.687165e-02
#>    [18,] -6.217043e-04  2.348610e-02  7.842285e-02
#>    [19,] -3.195898e-04 -1.314877e-01  2.500453e-02
#>    [20,]  3.932277e-04  7.492543e-02 -2.273731e-02
#>    [21,]  3.505620e-04 -1.794866e-02 -1.806379e-01
#>    [22,]  1.102100e-04  1.164253e-01 -2.508056e-02
#>    [23,] -3.179465e-04 -7.579651e-02 -1.099553e-01
#>    [24,]  4.133388e-06 -9.654133e-02 -3.805130e-02
#>    [25,] -3.054593e-04 -5.669092e-02 -1.153967e-01
#>    [26,] -1.268943e-03  4.365102e-02 -4.745854e-02
#>    [27,] -1.087195e-03 -8.392087e-02 -6.978389e-02
#>    [28,] -5.171225e-04 -9.190345e-02  4.600776e-02
#>    [29,] -1.127023e-03 -8.319932e-02  5.169016e-02
#>    [30,] -1.120905e-03 -1.805355e-02  7.011475e-02
#>    [31,] -6.047736e-04  1.474651e-02  2.377479e-02
#>    [32,] -1.517487e-03 -1.064388e-02 -4.558984e-02
#>    [33,] -2.789827e-04 -9.260273e-02 -7.429035e-02
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>               tau_1      tau_2      tau_3      tau_4      tau_5      tau_6
#>               tau_7      tau_8      tau_9      tau_10     tau_11     tau_12
#>             tau_13    tau_14     tau_15    tau_16    tau_17     tau_18
#>              tau_19     tau_20     tau_21    tau_22     tau_23    tau_24
#>              tau_25     tau_26     tau_27     tau_28     tau_29    tau_30
#>              tau_31     tau_32     tau_33    tau_34     tau_35     tau_36
#>              tau_37     tau_38    tau_39     tau_40     tau_41    tau_42
#>              tau_43     tau_44     tau_45     tau_46     tau_47    tau_48
#>               tau_49     tau_50     tau_51     tau_52    tau_53     tau_54
#>             tau_55     tau_56     tau_57    tau_58    tau_59     tau_60
#>              tau_61     tau_62    tau_63      tau_64    tau_65    tau_66
#>              tau_67     tau_68    tau_69     tau_70     tau_71    tau_72
#>              tau_73     tau_74    tau_75     tau_76     tau_77     tau_78
#>              tau_79    tau_80     tau_81    tau_82     tau_83     tau_84
#>              tau_85     tau_86    tau_87    tau_88     tau_89     tau_90
#>              tau_91     tau_92     tau_93     tau_94    tau_95     tau_96
#>              tau_97     tau_98     tau_99    tau_100    tau_101    tau_102
#>             tau_103   tau_104    tau_105    tau_106    tau_107    tau_108
#>             tau_109    tau_110    tau_111    tau_112    tau_113     tau_114
#>             tau_115    tau_116    tau_117    tau_118   tau_119    tau_120
#>             tau_121    tau_122    tau_123    tau_124   tau_125    tau_126
#>             tau_127    tau_128    tau_129   tau_130    tau_131   tau_132
#>            tau_133    tau_134    tau_135    tau_136    tau_137    tau_138
#>             tau_139    tau_140    tau_141    tau_142    tau_143    tau_144
#>             tau_145    tau_146    tau_147    tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }