Common Extractors for 'svdraws' and 'svpredict' Objects

Some simple extractors returning the corresponding element of an svdraws and svpredict object.

para(x, chain = "concatenated")

latent0(x, chain = "concatenated")

latent(x, chain = "concatenated")

vola(x, chain = "concatenated")

svbeta(x, chain = "concatenated")

svtau(x, chain = "concatenated")

priors(x)

thinning(x)

runtime(x)

sampled_parameters(x)

predy(y, chain = "concatenated")

predlatent(y, chain = "concatenated")

predvola(y, chain = "concatenated")

Arguments

x

svdraws object.

chain

optional either a positive integer or the string "concatenated" (default) or the string "all".

y

svpredict object.

Value

The return value depends on the actual funtion.

para(x, chain = "concatenated")

extracts the parameter draws.

latent(x, chain = "concatenated")

extracts the latent contemporaneous log-volatility draws.

latent0(x, chain = "concatenated")

extracts the latent initial log-volatility draws.

svbeta(x, chain = "concatenated")

extracts the linear regression coefficient draws.

svtau(x, chain = "concatenated")

extracts the tau draws.

vola(x, chain = "concatenated")

extracts standard deviation draws.

priors(x)

extracts the prior parameters used and returns them in a prior_spec object as generated by specify_priors.

thinning(x)

extracts the thinning parameters used and returns them in a list.

runtime(x)

extracts the runtime and returns it as a proc_time object.

sampled_parameters(x)

returns the names of time independent model parameters that were actually sampled by svsample.

predlatent(y, chain = "concatenated")

extracts the predicted latent contemporaneous log-volatility draws.

predvola(y, chain = "concatenated")

extracts predicted standard deviation draws.

predy(y, chain = "concatenated")

extracts the predicted observation draws.

Functions that have input parameter chain return an mcmc.list object if chain=="all" and return an mcmc object otherwise. If chain is an integer, then the specified chain is selected from all chains. If chain is "concatenated", then all chains are merged into one mcmc object.

See also

Examples

# Simulate data
sim <- svsim(150)

# Draw from vanilla SV
draws <- svsample(sim, draws = 2000)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Summarize all estimated parameter draws as a merged mcmc object
summary(para(draws)[, sampled_parameters(draws)])
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> mu    -9.7830 0.6421 0.014358        0.03073
#> phi    0.8693 0.1848 0.004133        0.03900
#> sigma  0.2475 0.1223 0.002734        0.01709
#> 
#> 2. Quantiles for each variable:
#> 
#>            2.5%     25%     50%     75%   97.5%
#> mu    -10.97344 -9.9570 -9.7259 -9.5590 -8.9480
#> phi     0.27104  0.8511  0.9252  0.9666  0.9931
#> sigma   0.07994  0.1602  0.2234  0.3093  0.5494
#> 
## Extract the draws as an mcmc.list object
params <- para(draws, chain = "all")[, sampled_parameters(draws)]

# \donttest{
options(max.print = 100)
## Further short examples
summary(latent0(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean             SD       Naive SE Time-series SE 
#>       -9.71848        0.50040        0.01119        0.01880 
#> 
#> 2. Quantiles for each variable:
#> 
#>    2.5%     25%     50%     75%   97.5% 
#> -10.683 -10.041  -9.719  -9.419  -8.675 
#> 
summary(latent(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> h_1    -9.717 0.4633 0.010360        0.01971
#> h_2    -9.767 0.4647 0.010390        0.02013
#> h_3    -9.813 0.4833 0.010806        0.02268
#> h_4    -9.823 0.4560 0.010197        0.02226
#> h_5    -9.804 0.4411 0.009864        0.02199
#> h_6    -9.791 0.4399 0.009836        0.02170
#> h_7    -9.775 0.4435 0.009918        0.02114
#> h_8    -9.713 0.4136 0.009248        0.02103
#> h_9    -9.681 0.4045 0.009045        0.02055
#> h_10   -9.690 0.4141 0.009260        0.01930
#> h_11   -9.651 0.3973 0.008883        0.01859
#> h_12   -9.684 0.4116 0.009204        0.01763
#> h_13   -9.678 0.3994 0.008931        0.01716
#> h_14   -9.692 0.4099 0.009166        0.01920
#> h_15   -9.650 0.4011 0.008968        0.01767
#> h_16   -9.590 0.3834 0.008572        0.02212
#> h_17   -9.644 0.4088 0.009141        0.01644
#> h_18   -9.658 0.4008 0.008963        0.01663
#> h_19   -9.739 0.4265 0.009538        0.01735
#> h_20   -9.769 0.4526 0.010121        0.02069
#> h_21   -9.781 0.4469 0.009993        0.02138
#> h_22   -9.783 0.4394 0.009826        0.02066
#> h_23   -9.787 0.4509 0.010083        0.02195
#> h_24   -9.752 0.4344 0.009713        0.02188
#> h_25   -9.779 0.4455 0.009962        0.02205
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>          2.5%     25%     50%     75%  97.5%
#> h_1   -10.594 -10.014  -9.712  -9.420 -8.765
#> h_2   -10.717 -10.074  -9.752  -9.467 -8.848
#> h_3   -10.830 -10.095  -9.791  -9.518 -8.889
#> h_4   -10.752 -10.120  -9.800  -9.524 -8.929
#> h_5   -10.694 -10.084  -9.793  -9.518 -8.930
#> h_6   -10.675 -10.067  -9.764  -9.507 -8.931
#> h_7   -10.696 -10.051  -9.752  -9.487 -8.915
#> h_8   -10.533  -9.984  -9.696  -9.455 -8.870
#> h_9   -10.473  -9.952  -9.666  -9.418 -8.856
#> h_10  -10.510  -9.955  -9.670  -9.429 -8.875
#> h_11  -10.420  -9.909  -9.648  -9.407 -8.868
#> h_12  -10.500  -9.951  -9.668  -9.421 -8.870
#> h_13  -10.460  -9.940  -9.672  -9.423 -8.867
#> h_14  -10.534  -9.947  -9.678  -9.429 -8.871
#> h_15  -10.447  -9.906  -9.635  -9.398 -8.831
#> h_16  -10.309  -9.840  -9.592  -9.367 -8.733
#> h_17  -10.413  -9.907  -9.647  -9.399 -8.792
#> h_18  -10.435  -9.917  -9.659  -9.416 -8.858
#> h_19  -10.625  -9.998  -9.731  -9.465 -8.908
#> h_20  -10.734 -10.046  -9.748  -9.478 -8.911
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
summary(vola(draws))
#> 
#> Iterations = 1001:3000
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 2000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>           Mean       SD  Naive SE Time-series SE
#> h_1   0.007975 0.001921 4.295e-05      7.927e-05
#> h_2   0.007777 0.001862 4.164e-05      7.954e-05
#> h_3   0.007614 0.001863 4.167e-05      8.339e-05
#> h_4   0.007555 0.001733 3.875e-05      8.362e-05
#> h_5   0.007613 0.001691 3.782e-05      8.204e-05
#> h_6   0.007661 0.001688 3.775e-05      8.630e-05
#> h_7   0.007727 0.001707 3.817e-05      7.872e-05
#> h_8   0.007947 0.001674 3.744e-05      8.223e-05
#> h_9   0.008066 0.001662 3.716e-05      8.451e-05
#> h_10  0.008038 0.001691 3.781e-05      7.820e-05
#> h_11  0.008183 0.001669 3.733e-05      7.738e-05
#> h_12  0.008062 0.001695 3.791e-05      7.114e-05
#> h_13  0.008075 0.001651 3.692e-05      6.897e-05
#> h_14  0.008026 0.001670 3.735e-05      7.354e-05
#> h_15  0.008191 0.001672 3.739e-05      7.539e-05
#> h_16  0.008428 0.001691 3.782e-05      9.708e-05
#> h_17  0.008225 0.001762 3.939e-05      6.994e-05
#> h_18  0.008157 0.001694 3.787e-05      6.641e-05
#> h_19  0.007854 0.001685 3.768e-05      6.625e-05
#> h_20  0.007757 0.001744 3.899e-05      7.683e-05
#> h_21  0.007705 0.001715 3.835e-05      7.633e-05
#> h_22  0.007689 0.001667 3.727e-05      8.405e-05
#> h_23  0.007684 0.001719 3.844e-05      8.337e-05
#> h_24  0.007809 0.001694 3.788e-05      8.223e-05
#> h_25  0.007711 0.001679 3.754e-05      8.045e-05
#>  [ reached getOption("max.print") -- omitted 125 rows ]
#> 
#> 2. Quantiles for each variable:
#> 
#>           2.5%      25%      50%      75%    97.5%
#> h_1   0.005006 0.006690 0.007781 0.009004 0.012491
#> h_2   0.004708 0.006492 0.007627 0.008796 0.011986
#> h_3   0.004450 0.006426 0.007480 0.008575 0.011740
#> h_4   0.004626 0.006345 0.007446 0.008550 0.011508
#> h_5   0.004762 0.006460 0.007474 0.008574 0.011503
#> h_6   0.004808 0.006516 0.007582 0.008622 0.011499
#> h_7   0.004758 0.006567 0.007629 0.008709 0.011590
#> h_8   0.005160 0.006793 0.007842 0.008847 0.011854
#> h_9   0.005319 0.006902 0.007963 0.009012 0.011940
#> h_10  0.005222 0.006890 0.007945 0.008965 0.011824
#> h_11  0.005462 0.007050 0.008036 0.009062 0.011867
#> h_12  0.005246 0.006905 0.007955 0.009000 0.011853
#> h_13  0.005352 0.006944 0.007938 0.008992 0.011872
#> h_14  0.005158 0.006919 0.007916 0.008962 0.011849
#> h_15  0.005390 0.007062 0.008086 0.009106 0.012091
#> h_16  0.005773 0.007299 0.008262 0.009249 0.012692
#> h_17  0.005482 0.007060 0.008038 0.009101 0.012329
#> h_18  0.005422 0.007022 0.007989 0.009025 0.011928
#> h_19  0.004931 0.006743 0.007707 0.008803 0.011631
#> h_20  0.004667 0.006585 0.007644 0.008748 0.011615
#>  [ reached getOption("max.print") -- omitted 130 rows ]
#> 
sampled_parameters(draws)
#> [1] "mu"    "phi"   "sigma"
priors(draws)
#> Prior distributions:
#> mu        ~ Normal(mean = 0, sd = 100)
#> (phi+1)/2 ~ Beta(a = 5, b = 1.5)
#> sigma^2   ~ Gamma(shape = 0.5, rate = 0.5)
#> nu        ~ Infinity
#> rho       ~ Constant(value = 0)

# Draw 3 independent chains from heavy-tailed and asymmetric SV with AR(2) structure
draws <- svsample(sim, draws = 20000, burnin = 3000,
                  designmatrix = "ar2",
                  priornu = 0.1, priorrho = c(4, 4),
                  n_chains = 3)
#> Extracted data vector from 'svsim'-object.
#> Done!
#> Summarizing posterior draws...

## Extract beta draws from the second chain
svbeta(draws, chain = 2)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>                 beta_0        beta_1        beta_2
#>     [1,]  4.234132e-04  3.217441e-02 -6.636162e-02
#>     [2,]  3.992005e-04  2.277953e-02 -8.173397e-02
#>     [3,]  4.057915e-04  9.371802e-02 -8.814401e-02
#>     [4,]  9.935080e-04  2.065975e-01 -6.067773e-02
#>     [5,] -9.609262e-04  1.692536e-01 -1.886064e-01
#>     [6,] -1.460520e-04  5.145940e-02 -9.538988e-02
#>     [7,] -8.973373e-04  2.687983e-02 -1.658402e-01
#>     [8,] -5.772455e-04 -9.878886e-02 -9.804015e-02
#>     [9,]  5.769360e-04  1.651826e-01 -2.526125e-01
#>    [10,] -1.566845e-04  1.495848e-01 -2.077190e-01
#>    [11,] -1.050361e-03  6.089830e-02 -2.592425e-01
#>    [12,] -9.410727e-04  1.459953e-01 -1.522750e-01
#>    [13,] -1.575821e-03  7.331976e-02 -2.315218e-01
#>    [14,] -1.435312e-03  7.470032e-02 -1.665854e-01
#>    [15,] -4.749403e-04 -8.067770e-02 -1.936039e-01
#>    [16,] -2.300189e-03  5.857767e-02 -6.305124e-02
#>    [17,] -5.146535e-04  3.750542e-02 -1.144795e-01
#>    [18,] -3.216412e-04  6.210668e-02 -2.047618e-01
#>    [19,] -1.063469e-03 -1.817913e-01 -2.572698e-01
#>    [20,]  3.228282e-04  1.007802e-01 -2.357437e-01
#>    [21,]  3.322229e-04  6.756113e-02 -1.459565e-01
#>    [22,] -7.818805e-04 -5.345473e-02 -1.445120e-01
#>    [23,] -7.412510e-04 -3.409482e-02 -1.072822e-01
#>    [24,] -3.423187e-04  1.468712e-01 -4.546589e-02
#>    [25,]  6.826695e-04  1.111358e-01 -7.366736e-02
#>    [26,]  2.364554e-04  1.018870e-01 -1.351474e-01
#>    [27,] -5.877291e-04  1.317870e-01 -6.919620e-02
#>    [28,] -2.833365e-04  9.816123e-02 -1.523701e-01
#>    [29,]  7.618187e-05  8.050415e-02 -3.241153e-01
#>    [30,] -6.617500e-04 -9.262085e-02 -1.170962e-01
#>    [31,]  1.684366e-04  2.818824e-02  1.752888e-02
#>    [32,]  1.434995e-03  6.446379e-02 -6.145356e-02
#>    [33,] -2.234336e-04  1.349889e-01 -1.047744e-01
#>  [ reached getOption("max.print") -- omitted 19967 rows ]
## ... tau draws from all chains merged/concatenated together
svtau(draws)
#> Markov Chain Monte Carlo (MCMC) output:
#> Start = 3001 
#> End = 23000 
#> Thinning interval = 1 
#>               tau_1      tau_2     tau_3      tau_4      tau_5     tau_6
#>               tau_7      tau_8     tau_9     tau_10    tau_11     tau_12
#>             tau_13     tau_14    tau_15     tau_16     tau_17     tau_18
#>             tau_19     tau_20     tau_21     tau_22    tau_23     tau_24
#>              tau_25    tau_26    tau_27     tau_28    tau_29     tau_30
#>             tau_31    tau_32    tau_33    tau_34    tau_35     tau_36    tau_37
#>             tau_38     tau_39     tau_40     tau_41     tau_42    tau_43
#>              tau_44     tau_45    tau_46     tau_47     tau_48    tau_49
#>             tau_50    tau_51    tau_52     tau_53     tau_54     tau_55
#>             tau_56    tau_57    tau_58    tau_59     tau_60    tau_61    tau_62
#>              tau_63     tau_64    tau_65    tau_66    tau_67    tau_68
#>             tau_69    tau_70    tau_71    tau_72    tau_73    tau_74    tau_75
#>              tau_76    tau_77    tau_78     tau_79    tau_80    tau_81
#>             tau_82    tau_83     tau_84     tau_85    tau_86    tau_87
#>              tau_88     tau_89     tau_90    tau_91    tau_92     tau_93
#>              tau_94    tau_95    tau_96     tau_97    tau_98    tau_99
#>            tau_100   tau_101    tau_102    tau_103   tau_104   tau_105
#>            tau_106    tau_107    tau_108   tau_109   tau_110   tau_111
#>            tau_112    tau_113   tau_114    tau_115    tau_116    tau_117
#>            tau_118   tau_119   tau_120   tau_121    tau_122     tau_123
#>             tau_124   tau_125   tau_126    tau_127    tau_128   tau_129
#>             tau_130    tau_131   tau_132   tau_133   tau_134   tau_135
#>             tau_136   tau_137    tau_138   tau_139   tau_140   tau_141
#>            tau_142    tau_143    tau_144    tau_145   tau_146   tau_147
#>            tau_148
#>  [ reached getOption("max.print") -- omitted 20000 rows ]
## Create a new svdraws object from the first and third chain
second_chain_excluded <- draws[c(1, 3)]

# Draw from the predictive distribution
pred <- predict(draws, steps = 2)

## Extract the predicted observations as an mcmc.list object
predicted_y <- predy(pred, chain = "all")
## ... the predicted standard deviations from the second chain
predicted_sd <- predvola(pred, chain = 2)
## Create a new svpredict object from the first and third chain
second_chain_excluded <- pred[c(1, 3)]
# }